% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{add_protein_prettyprint_label}
\alias{add_protein_prettyprint_label}
\title{add short prettyprint labels to a protein tibble using gene symbols, if available}
\usage{
add_protein_prettyprint_label(tib, maxlength = 12, shorten_ambiguous = TRUE)
}
\arguments{
\item{tib}{tibble with at least a protein_id column, but should also contain a gene_symbols_or_id column for truly pretty-print names}

\item{maxlength}{max string length for each output value}

\item{shorten_ambiguous}{replace all ambiguous entries with the first value followed by an asterix. Example; \verb{proteinA;proteinB} becomes  \verb{proteinA*}}
}
\value{
input tibble with resulting values in the label column (overwritten if this column already existed in input)
}
\description{
add short prettyprint labels to a protein tibble using gene symbols, if available
}
