% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{adjust_modes}
\alias{adjust_modes}
\title{scale sample k by scaling factor s_k  -->>  don't have to re-compute all scalings, can just update the pairwise matrix}
\usage{
adjust_modes(m, s)
}
\arguments{
\item{m}{matrix with pairwise foldchange modes}

\item{s}{array of scaling factors for each sample}
}
\value{
updated matrix m
}
\description{
scale sample k by scaling factor s_k  -->>  don't have to re-compute all scalings, can just update the pairwise matrix
}
