% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{aggregate_by_group}
\alias{aggregate_by_group}
\title{placeholder title}
\usage{
aggregate_by_group(mat, groups, FUN, ...)
}
\arguments{
\item{mat}{todo}

\item{groups}{todo}

\item{FUN}{todo}

\item{...}{todo}
}
\description{
placeholder title
}
