% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{aggregate_tibble_by_datatables}
\alias{aggregate_tibble_by_datatables}
\title{highly optimized aggregation of the peptides data table.
rewrite aggregate_tibble_by_datatables; more speed, support charge/mz columns (if present), take RT from most abundant peptide with prio for detected. the latter matters when merging data for plain sequences where one precursor has an MS/MS identification and the other is MBR}
\usage{
aggregate_tibble_by_datatables(peptides, prop_peptide = "sequence_plain")
}
\arguments{
\item{peptides}{todo}

\item{prop_peptide}{must be either 'sequence_plain', 'sequence_modified' or 'peptide_id' (the former 2 are the common use-case)}
}
\description{
highly optimized aggregation of the peptides data table.
rewrite aggregate_tibble_by_datatables; more speed, support charge/mz columns (if present), take RT from most abundant peptide with prio for detected. the latter matters when merging data for plain sequences where one precursor has an MS/MS identification and the other is MBR
}
