% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{append_log_error}
\alias{append_log_error}
\title{Append error object to log}
\usage{
append_log_error(err, type = "error")
}
\arguments{
\item{err}{error type object (e.g. from tryCatch function)}

\item{type}{the type of message, typically "progress". See further append_log()}
}
\description{
prints message and call trace if these exist, otherwise dump the entire error object
}
