% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{as_matrix_except_first_column}
\alias{as_matrix_except_first_column}
\title{placeholder title}
\usage{
as_matrix_except_first_column(x)
}
\arguments{
\item{x}{todo}
}
\description{
placeholder title
}
