% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{assert_lme4_functional}
\alias{assert_lme4_functional}
\title{Throw an error if lme4::lmer() cannot be executed}
\usage{
assert_lme4_functional()
}
\description{
\code{stop()} is called when lme4 or Matrix is not installed, or when \code{lme4::lmer()}
cannot be executed (with default settings and mock data).
If this function does not throw an error, we may assume \code{lme4::lmer()} is operational.
}
