% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{boxplot_add_text}
\alias{boxplot_add_text}
\title{placeholder title}
\usage{
boxplot_add_text(bp, show_N = TRUE, show_values = T, ...)
}
\arguments{
\item{bp}{todo}

\item{show_N}{todo}

\item{show_values}{todo}

\item{...}{todo}
}
\description{
placeholder title
}
