% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{cache_filtering_data}
\alias{cache_filtering_data}
\title{Cache data required for fast downstream peptide filtering}
\usage{
cache_filtering_data(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
eg; number of replicate samples per group in which a peptide is detected, within-group variation, etc.
}
