% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{check_dataset_hascache}
\alias{check_dataset_hascache}
\title{check if peptide tibble has cache}
\usage{
check_dataset_hascache(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
check if peptide tibble has cache
}
