% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_dataset_integrity}
\alias{check_dataset_integrity}
\title{Test if the minimum set of required information is present in a dataset}
\usage{
check_dataset_integrity(dataset)
}
\arguments{
\item{dataset}{dataset to validate}
}
\description{
Test if the minimum set of required information is present in a dataset
}
