% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{check_parameter_is_boolean}
\alias{check_parameter_is_boolean}
\title{generic input validation; all provided arguments should be boolean/logical (not array)}
\usage{
check_parameter_is_boolean(...)
}
\arguments{
\item{...}{arbitrary set of parameters}
}
\description{
generic input validation; all provided arguments should be boolean/logical (not array)
}
