% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{check_parameter_is_numeric}
\alias{check_parameter_is_numeric}
\title{generic input validation; all provided arguments should be numeric (not array)}
\usage{
check_parameter_is_numeric(..., minval_ = NA, maxval_ = NA)
}
\arguments{
\item{...}{arbitrary set of parameters}

\item{minval_}{optionally, a finite number indicating the minimum value for each parameter}

\item{maxval_}{optionally, a finite number indicating the maximum value for each parameter}
}
\description{
generic input validation; all provided arguments should be numeric (not array)
}
