% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{check_parameter_is_string}
\alias{check_parameter_is_string}
\title{generic input validation; all provided arguments should be strings (not array)}
\usage{
check_parameter_is_string(...)
}
\arguments{
\item{...}{arbitrary set of parameters}
}
\description{
generic input validation; all provided arguments should be strings (not array)
}
