% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_valid_tibble_peptides}
\alias{check_valid_tibble_peptides}
\title{data integrity checks for peptide tibble}
\usage{
check_valid_tibble_peptides(tib)
}
\arguments{
\item{tib}{peptide tibble (eg; dataset$peptides)}
}
\description{
data integrity checks for peptide tibble
}
