% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_valid_tibble_proteins}
\alias{check_valid_tibble_proteins}
\title{data integrity checks for protein tibble}
\usage{
check_valid_tibble_proteins(tib)
}
\arguments{
\item{tib}{protein tibble (eg; dataset$proteins)}
}
\description{
data integrity checks for protein tibble
}
