% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{check_valid_tibble_samples}
\alias{check_valid_tibble_samples}
\title{data integrity checks for sample tibble}
\usage{
check_valid_tibble_samples(tib)
}
\arguments{
\item{tib}{sample tibble (eg; dataset$samples)}
}
\description{
data integrity checks for sample tibble
}
