% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{coefficient_of_variation}
\alias{coefficient_of_variation}
\title{placeholder title}
\usage{
coefficient_of_variation(nonlog_array)
}
\arguments{
\item{nonlog_array}{todo}
}
\description{
placeholder title
}
