% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{coefficient_of_variation_vectorized}
\alias{coefficient_of_variation_vectorized}
\title{fast CoV computation}
\usage{
coefficient_of_variation_vectorized(matrix_with_naturallog_intensities)
}
\arguments{
\item{matrix_with_naturallog_intensities}{a matrix with values transformed into natural log}
}
\description{
fast CoV computation
}
