% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{column_intensity_to_label}
\alias{column_intensity_to_label}
\title{pretty-print label for an intensity column}
\usage{
column_intensity_to_label(col)
}
\arguments{
\item{col}{todo}
}
\description{
pretty-print label for an intensity column
}
