% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{contrast_to_samplegroups}
\alias{contrast_to_samplegroups}
\title{return sample groups for a given contrast (column name in dataset$samples)}
\usage{
contrast_to_samplegroups(dataset, contr_lbl)
}
\arguments{
\item{dataset}{your dataset}

\item{contr_lbl}{a string describing a contrast = column in dataset$samples, e.g. "contrast: WT vs KO"}
}
\value{
list of unique sample groups, first element is left-hand side of the contrast and second element the right-hand side
}
\description{
return sample groups for a given contrast (column name in dataset$samples)
}
