% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{count_sep_char}
\alias{count_sep_char}
\title{Count how often commonly used separation characters occur in an array of strings}
\usage{
count_sep_char(x)
}
\description{
Count how often commonly used separation characters occur in an array of strings
}
