% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_contrasts}
\alias{dataset_contrasts}
\title{List the name of all contrasts in the samples table}
\usage{
dataset_contrasts(dataset)
}
\arguments{
\item{dataset}{a valid dataset}
}
\description{
List the name of all contrasts in the samples table
}
