% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{dataset_transfer_peptide_properties}
\alias{dataset_transfer_peptide_properties}
\title{carries over protein_id, sequence_plain and sequence_modified in rows where protein_id is NA}
\usage{
dataset_transfer_peptide_properties(dataset)
}
\arguments{
\item{dataset}{input dataset}
}
\description{
all NA values in detect and isdecoy columns will be set to FALSE
}
