% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_interface_ebayes}
\alias{de_interface_ebayes}
\title{placeholder title}
\usage{
de_interface_ebayes(
  peptides = NULL,
  samples = NULL,
  eset_peptides = NULL,
  eset_proteins = NULL,
  input_intensities_are_log2 = T
)
}
\arguments{
\item{peptides}{todo}

\item{samples}{todo}

\item{eset_peptides}{todo}

\item{eset_proteins}{todo}

\item{input_intensities_are_log2}{todo}
}
\description{
placeholder title
}
