% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_limpa}
\alias{de_limpa}
\title{Apply limpa}
\usage{
de_limpa(
  eset,
  model_matrix,
  model_matrix_result_prop,
  doplot = FALSE,
  ebayes_trend = FALSE,
  ebayes_robust = FALSE
)
}
\arguments{
\item{eset}{protein-level log2 intensity matrix stored as a Biobase ExpressionSet}

\item{model_matrix}{a \code{stats::model.matrix()} result that is supplied to \code{limma::lmFit()}}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}

\item{doplot}{not implemented yet}

\item{ebayes_trend}{boolean parameter to set the \code{trend} parameter in \code{limma::eBayes()}. Default: FALSE. Alternatively, use \code{options(msdap.dea.ebayes.trend = TRUE)}}

\item{ebayes_robust}{boolean parameter to set the \code{robust} parameter in \code{limma::eBayes()}. Default: FALSE. Alternatively, use \code{options(msdap.dea.ebayes.robust = TRUE)}}
}
\description{
ref; Li M, Cobbold SA, Smyth GK (2025). “Quantification and differential analysis of mass spectrometry proteomics data with probabilistic recovery of information from missing values.” bioRxiv, April 29 2025, 651125. doi:10.1101/2025.04.28.651125
ref; https://bioconductor.org/packages/release/bioc/html/limpa.html
}
\details{
This implementation follows the example from; https://bioconductor.org/packages/release/bioc/vignettes/limpa/inst/doc/limpa.html
}
