% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msempire}
\alias{de_msempire}
\title{MS-EmpiRe implementation, a wrapper function for msEmpiRe::de.ana()}
\usage{
de_msempire(eset, model_matrix, model_matrix_result_prop)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the peptide log2 intensities}

\item{model_matrix}{a \code{stats::model.matrix()} result that is supplied to \code{limma::lmFit()}}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}
}
\description{
ref; PMID:31235637
ref; https://github.com/zimmerlab/MS-EmpiRe
}
\details{
note; MS-EmpiRe crashes when all proteins have the same number of peptides (eg; filtering topn=1 or topn=3 and minpep=3)

TODO: proper standarderror computation
}
