% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msqrob_customformula}
\alias{de_msqrob_customformula}
\title{MSqRob test function for internal use}
\usage{
de_msqrob_customformula(
  peptides,
  proteins,
  samples,
  main_variable = "condition",
  ordered_variables = c("batch", "condition"),
  min_values = NULL
)
}
\arguments{
\item{peptides}{TODO}

\item{proteins}{TODO}

\item{samples}{TODO}

\item{main_variable}{TODO}

\item{ordered_variables}{TODO}

\item{min_values}{minimum number of data points in each condition. Alternatively, use \code{options(msdap.dea.minvalue = 5)}}
}
\description{
MSqRob test function for internal use
}
