% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msqrobsum_msqrob}
\alias{de_msqrobsum_msqrob}
\title{Wrapper function for msqrob, using the implementation from the msqrobsum package}
\usage{
de_msqrobsum_msqrob(
  eset,
  model_matrix,
  model_matrix_result_prop,
  use_peptide_model,
  random_variables = NULL
)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the peptide log2 intensities}

\item{model_matrix}{a \code{stats::model.matrix()} result that is supplied to \code{limma::lmFit()}}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}

\item{use_peptide_model}{if \code{TRUE}, apply msqrob. if \code{FALSE}, apply msqrobsum}

\item{random_variables}{optionally, an array of column names in the sample metadata table that should be used as additional regression variables}
}
\description{
note that this is the final version of this algorithm provided by its authors, and is no longer maintained; https://github.com/statOmics/msqrob
only difference with the GitHub version is that code included here has minor adaptions to situationally improve multithreading (i.e. speeds up computation a bit, on some systems)
}
\details{
ref; PMID:26566788 PMID:32321741
ref; https://github.com/statOmics/msqrob
}
