% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea}
\alias{dea}
\title{Differential expression analysis}
\usage{
dea(
  dataset,
  qval_signif = 0.05,
  fc_signif = 0,
  algo_de = c("ebayes"),
  output_dir_for_eset = ""
)
}
\arguments{
\item{dataset}{your dataset}

\item{qval_signif}{threshold for significance of adjusted p-values. default: 0.05}

\item{fc_signif}{threshold for significance of log2 foldchanges. Set to zero to disregard, a positive value to apply a cutoff to absolute foldchanges or use bootstrap analyses to infer a suitable foldchange threshold by providing either NA or a negative value. default: 0}

\item{algo_de}{algorithms for differential expression analysis. options: ebayes, msempire, msqrob (to run multiple, provide an array)}

\item{output_dir_for_eset}{optionally, provide an output directory where the expressionset objects should be stored. Only useful if you're doing downstream data analysis that required this data}
}
\description{
Differential expression analysis
}
