% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_algorithms}
\alias{dea_algorithms}
\title{returns all DEA functions integrated with MS-DAP}
\usage{
dea_algorithms()
}
\description{
\subsection{available DEA functions}{

\strong{ebayes}: wrapper for the eBayes function from the limma package (PMID:25605792) \url{https://bioconductor.org/packages/release/bioc/html/limma.html}. The eBayes function applies moderated t-statistics to each row of a given data matrix.
It was originally developed for the analysis of RNA-sequencing data but can also be applied to a proteomics \code{protein*sample} data matrix. Doesn't work on peptide-level data because limma eBayes returns t-statistics per row in the input matrix,
so using peptide-level data would yield statistics per peptide and translating those to protein-level statistics is not straight forward.
Thus, with MS-DAP we first perform peptide-to-protein rollup (e.g. using MaxLFQ algorithm) and then apply the limma eBayes function to the protein-level data matrix.

This is in line with typical usage of moderated t-statistics in proteomics (e.g. analogous to Perseus, where a moderated t-test is applied to protein data matrix). This method will take provided covariates into account (if any). Implemented in function; \code{de_ebayes}

\strong{deqms}: wrapper for the DEqMS package (PMID:32205417) \url{https://github.com/yafeng/DEqMS}, which is a proteomics-focussed extension of the limma eBayes approach that also weighs the number of peptides observed per protein to adjust protein statistics. MS-DAP will apply this function to the protein-level data matrix. This method will take provided covariates into account (if any). Implemented in function; \code{de_deqms}

\strong{msempire}: wrapper for the msEmpiRe package (PMID:31235637) \url{https://github.com/zimmerlab/MS-EmpiRe}. This is a peptide-level DEA algorithm. Note that this method cannot deal with covariates! Implemented in function; \code{de_msempire}

\strong{msqrob}: implementation of the MSqRob package, with minor tweak for (situationally) faster computation (PMID:26566788) \url{https://github.com/statOmics/msqrob}. This is a peptide-level DEA algorithm. This method will take provided covariates into account (if any). Implemented in function; \code{de_msqrobsum_msqrob}

\strong{msqrobsum}: implementation of the MSqRob package (which also features MSqRobSum), with minor tweak for (situationally) faster computation (PMID:32321741) \url{https://github.com/statOmics/msqrob}. This is a hybrid peptide&protein-level DEA algorithm that takes peptide-level data as input; it first performs peptide-to-protein rollup, then applies statistics to this protein-level data matrix. This method will take provided covariates into account (if any). Implemented in function; \code{de_msqrobsum_msqrob}

\strong{limpa}: implementation of the limpa package; Li M, Cobbold SA, Smyth GK (2025). “Quantification and differential analysis of mass spectrometry proteomics data with probabilistic recovery of information from missing values.” bioRxiv, April 29 2025, 651125. doi:10.1101/2025.04.28.651125
}
}
