% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_algorithms_reserved}
\alias{dea_algorithms_reserved}
\title{reserved DEA function names}
\usage{
dea_algorithms_reserved()
}
\description{
The returned function names are reserved; custom DEA functions cannot have any of these names
}
