% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_protein_background_foldchange_limits}
\alias{dea_protein_background_foldchange_limits}
\title{estimate a threshold for 'significant' foldchanges from N permutations of sample-to-condition assignments}
\usage{
dea_protein_background_foldchange_limits(
  x,
  samples_cond1 = NULL,
  samples_cond2 = NULL,
  probs = 0.95,
  max_permutations = 100
)
}
\arguments{
\item{x}{a numeric matrix or a Biobase::ExpressionSet(). If the latter is provided, the ExpressionSet must contain a column named "condition" in phenotypic data and it must contain numeric values (which will be casted to integers), where 1 indicates condition 1 and 2 indicates condition 2}

\item{samples_cond1}{if x is a matrix, this must be a character vector indicating samples from condition 1; these must be column names in x}

\item{samples_cond2}{analogous to samples_cond1, but for condition 2}

\item{probs}{upper limit for the quantile cutoff, automatically translated to mirror the lower limit; c(1-probs, probs)}

\item{max_permutations}{maximum number of unique configurations used for the permuted datasets}
}
\description{
note; this function hardcodes set.seed()

Permutations of sample labels within a group are disregarded as these have no effect on the between-group foldchange, only unique combinations of swapping samples between conditions A and B are considered

This is somewhat similar to the method described by Hafemeister and Satija at https://doi.org/10.1186/s13059-019-1874-1
M&M quote from this reference: "A random background distribution of mean differences was generated by randomly choosing 1000 genes and permuting the group labels. Significance thresholds for the difference of means were derived from the background distribution by taking the 0.5th and 99.5th percentile."
}
\examples{
\dontrun{
m = matrix(runif(600), nrow=100, ncol=6, dimnames=list(NULL, LETTERS[1:6]))
dea_protein_background_foldchange_limits__v2(
  m,
  samples_cond1 = c("A","B","C"), samples_cond2 = c("D","E","F"),
  probs = 0.95, max_permutations = 100
)
}
}
