% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_results_to_wide}
\alias{dea_results_to_wide}
\title{convert the results from differential expression analysis from a long-format tibble to wide-format}
\usage{
dea_results_to_wide(dataset)
}
\arguments{
\item{dataset}{your dataset. if 'de_proteins' is lacking, result is empty tibble}
}
\description{
DEA statistics summary table in wide format: protein_id, accessions, fasta_headers, gene_symbols_or_id, <algo_de x contrast>foldchange.log2, <algo_de x contrast>pvalue, etc.
if there are 3 or more different DEA algorithms in the results, add a column that combines their results such that all proteins significant in 2 or more tests/algorithms are flagged
}
