% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dea_summary_prettyprint}
\alias{dea_summary_prettyprint}
\title{prettyprint table that summarizes DEA results}
\usage{
dea_summary_prettyprint(dataset, trim_contrast_names = FALSE)
}
\arguments{
\item{dataset}{dataset that includes DD results}

\item{trim_contrast_names}{reduce string length of "A vs B" style contrast description}
}
\description{
prettyprint table that summarizes DEA results
}
