% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_detect.R
\name{diffdetect_results_to_wide}
\alias{diffdetect_results_to_wide}
\title{convert the results from differential detection analysis from a long-format tibble to wide-format}
\usage{
diffdetect_results_to_wide(dataset)
}
\arguments{
\item{dataset}{your dataset. if 'dd_proteins' is lacking, result is empty tibble}
}
\description{
convert the results from differential detection analysis from a long-format tibble to wide-format
}
