% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{diffdetect_summary_prettyprint}
\alias{diffdetect_summary_prettyprint}
\title{prettyprint table that summarizes differential detect results}
\usage{
diffdetect_summary_prettyprint(
  dataset,
  use_quant = FALSE,
  trim_contrast_names = FALSE
)
}
\arguments{
\item{dataset}{dataset that includes DD results}

\item{use_quant}{boolean indicating whether the 'detect' or 'quant' zcore results should be returned}

\item{trim_contrast_names}{reduce string length of "A vs B" style contrast description}
}
\description{
prettyprint table that summarizes differential detect results
}
