% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_detect.R
\name{differential_detect}
\alias{differential_detect}
\title{Compare the number of peptide detection counts between groups}
\usage{
differential_detect(
  dataset,
  min_peptides_observed = 1L,
  min_samples_observed = 3,
  min_fraction_observed = 0.5,
  count_mode = "auto",
  rescale_counts_per_sample = TRUE,
  return_wide_format = FALSE
)
}
\arguments{
\item{dataset}{a valid dataset}

\item{min_peptides_observed}{minimum number of peptides for a protein to pass filtering rules (i.e. otherwise, no z-score is computed)}

\item{min_samples_observed}{minimum number of samples where a protein should be observed with at least \code{min_peptides_observed} peptides (in either group) when comparing a contrast of group A vs B}

\item{min_fraction_observed}{for differential detection only; analogous to \code{diffdetect_min_samples_observed}, but here you can specify the fraction of samples where a protein needs to be detected in either group (within the respective contrast). default; 0.5 (50\% of samples)}

\item{count_mode}{whether z-scores for only detected peptides, or all observed peptides, should be computed. options; "auto", "detect", "quant"}

\item{rescale_counts_per_sample}{boolean, indicating whether per sample, detect counts should be rescaled (i.e. to normalize across samples for the number of detected peptides)}

\item{return_wide_format}{boolean, return table in simplified wide-format (TRUE) or return all data in long-format table (FALSE, default)}
}
\description{
The computation of scores is detailed in the online vignitte for "differential testing".
}
