% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{empty_protein_tibble}
\alias{empty_protein_tibble}
\title{placeholder long-format proteins tibble with minimum set of columns (protein_d, fasta_headers, gene_symbols_or_id) for some given peptides tibble}
\usage{
empty_protein_tibble(peptides)
}
\arguments{
\item{peptides}{peptide tibble in long format}
}
\description{
placeholder long-format proteins tibble with minimum set of columns (protein_d, fasta_headers, gene_symbols_or_id) for some given peptides tibble
}
