% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{enable_log}
\alias{enable_log}
\title{Enable or disable console logging}
\usage{
enable_log(enable = TRUE)
}
\arguments{
\item{enable}{boolean value; set to FALSE to suppress console logging while retaining the internal log buffer}
}
\description{
Enable or disable console logging
}
