% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{enforce_sample_value_types}
\alias{enforce_sample_value_types}
\title{convert the data type of column in a sample metadata table to factors or numeric type}
\usage{
enforce_sample_value_types(
  samples,
  redundant_columns = "error",
  show_log = TRUE
)
}
\arguments{
\item{samples}{dataset$samples or a subset thereof. Typically, only columns sample_id and user-defined regression variables are included as columns}

\item{redundant_columns}{how to deal with columns in \code{samples} that are redundant with other columns. Options: "error" (default), "warning", "ignore"}

\item{show_log}{boolean, show data type per column in console? Default: \code{TRUE}}
}
\description{
importantly, columns "sample_index", "sample_id" and "shortname" are ignored alltogether
}
