% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{error_legacy_contrast_definitions}
\alias{error_legacy_contrast_definitions}
\title{throw an error if has_legacy_contrast_definitions() is TRUE}
\usage{
error_legacy_contrast_definitions(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
throw an error if has_legacy_contrast_definitions() is TRUE
}
