% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{eset_from_peptides_to_proteins}
\alias{eset_from_peptides_to_proteins}
\title{Convert a peptide-level ExpressionSet to protein-level by summing their respective peptide intensities per sample}
\usage{
eset_from_peptides_to_proteins(eset_peptides)
}
\arguments{
\item{eset_peptides}{A Biobase ExpressionSet with peptide data. Intensity values are assumed to be log2 transformed! fData() must contain protein_id attributes}
}
\description{
Convert a peptide-level ExpressionSet to protein-level by summing their respective peptide intensities per sample
}
