% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_tables.R
\name{export_peptide_abundance_matrix}
\alias{export_peptide_abundance_matrix}
\title{Export peptide-level data to TSV files}
\usage{
export_peptide_abundance_matrix(dataset, output_dir)
}
\arguments{
\item{dataset}{your dataset}

\item{output_dir}{output directory where all output files are stored, must be an existing directory}
}
\description{
Generates separate files for input data as-is (i.e. all peptides, intensity values as provided in the imported dataset)
and all filtering&normalization applied to the dataset.
}
\details{
i.e. 1 output file for every column in dataset$peptides that starts with "intensity"
}
