% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_tables.R
\name{export_statistical_results}
\alias{export_statistical_results}
\title{write statistical results to file}
\usage{
export_statistical_results(dataset, output_dir)
}
\arguments{
\item{dataset}{your dataset}

\item{output_dir}{path to existing directory}
}
\description{
combines both the DEA and "differential testing" results into output file differential_abundance_analysis.xlsx
}
