% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fasta.R
\name{fasta_header_to_id}
\alias{fasta_header_to_id}
\title{extract the protein identifier from a fasta header}
\usage{
fasta_header_to_id(x)
}
\arguments{
\item{x}{array of fasta headers}
}
\description{
the first set of non-whitespace characters on the row is assumed to be the protein ID.
the leading '>' is not included (if present).
}
\details{
this regex should be robust for all sorts of input, including those not following official standards.
should be ">proteinid whatever". But if additional whitespace is provided at the start we are robust to this.

e.g. these should all yield 'pid' as the protein ID;
fasta_header_to_id(c(" > pid description", "> pid description", " pid description", "pid description"))
}
