% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fasta.R
\name{fasta_identifier_isdecoy}
\alias{fasta_identifier_isdecoy}
\title{regex test if a protein identifier or fasta header is a decoy}
\usage{
fasta_identifier_isdecoy(x)
}
\arguments{
\item{x}{strings to test}
}
\description{
regex test if a protein identifier or fasta header is a decoy
}
\examples{
\dontrun{
# decoy examples (should yield TRUE)
fasta_identifier_isdecoy(c(
">rev_sp|A0A023PZB3|FMP49_YEAST Protein FMP49, mitochondrial ...",
"rev_sp|A0A023PYF4|YE145_YEAST",
"decoy_A0A023PZF2",
"P63044_decoy",
"P63044_rev",
"P63044_reverse",
">sp|Q02956|KPCZ_MOUSE_decoy",
">decoy_sp|Q02956|KPCZ_MOUSE"
))
# non-decoy examples (should yield FALSE)
fasta_identifier_isdecoy(c(
">sp|A0A023PZB3|FMP49_YEAST Protein FMP49, mitochondrial ...",
"sp|A0A023PYF4|YE145_YEAST",
"A0A023PZF2",
"P63044",
">sp|Q02956|KPCZ_MOUSE"
))
}
}
