% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{file_check}
\alias{file_check}
\title{strip illegal characters from filename, then check for maximum path length}
\usage{
file_check(file, strict = FALSE)
}
\arguments{
\item{file}{filename to check (may include a path)}

\item{strict}{boolean indicating 'strict mode' for checking the filename, if TRUE; only allow alphanumeric, underscore and '-'}
}
\description{
If longer than N characters (Windows filesystem max path length), try with md5 hashed filename
}
\details{
note;
in contrast to using a timestamp or substring+index, this solution yields stable filenames. Should never fail if provided path/dir is reasonable
}
\examples{
\dontrun{
  x = c("C:/temp/file*.txt", "test=2", "test=2.abc",
         paste0("c:/temp/",paste(rep("a",300),collapse=""),".txt"))
  cbind(x, file_check(x))
}
\dontrun{
  file_check("C:/temp/") # error, trailing slashes not allowed
}
}
