% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{filename_extract_extension}
\alias{filename_extract_extension}
\title{extract the trailing file extension from a file or path}
\usage{
filename_extract_extension(f)
}
\arguments{
\item{f}{f can either be a filename or a full path, and optionally carry a file extension. eg; remove_file_extension_from_path("file.txt"); remove_file_extension_from_path("C:/temp/file")}
}
\value{
extension including the '.' or empty string if not found
}
\description{
purposely naive approach to keep things simple for our internal usecases; strip a . followed by 1-5 characters from end of the string
}
