% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{fit_t_dist_fixed_mu}
\alias{fit_t_dist_fixed_mu}
\title{use optim() to fit a t-distribution with fixed mu=0}
\usage{
fit_t_dist_fixed_mu(x)
}
\arguments{
\item{x}{numeric vector of at least 10 finite values}
}
\description{
use optim() to fit a t-distribution with fixed mu=0
}
\examples{
\dontrun{
# fit normal
x = rnorm(1000, sd = 2)
fit = suppressWarnings(fit_t_dist_fixed_mu(x))
h = hist(x, breaks = 25, freq = F)
curve(dt(x / fit[1], df=fit[2]) / fit[1], col = 2, add = T)
}
\dontrun{
# fit t-distribution
x = rt(1000, df = 3)
fit = suppressWarnings(fit_t_dist_fixed_mu(x))
h = hist(x, breaks = 25, freq = F)
curve(dt(x / fit[1], df=fit[2]) / fit[1], col = 2, add = T)
}
}
