% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{format_r_code}
\alias{format_r_code}
\title{format a string of R code using the formatR and styler packages}
\usage{
format_r_code(rcode_lines)
}
\arguments{
\item{rcode_lines}{R code as an array of characters (strings)}
}
\description{
parameter can either be a string or an array of strings (eg; lines from R history)
}
