% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_find_psm_files}
\alias{fragpipe_find_psm_files}
\title{Helper function to find psm.tsv files within a FragPipe output directory}
\usage{
fragpipe_find_psm_files(path)
}
\arguments{
\item{path}{FragPipe output directory}
}
\description{
File "filelist_ionquant.txt" contains a list of all psm.tsv files that are part of this dataset.
In older FragPipe versions, e.g. 15.0, this file is not available.

If filelist_ionquant.txt is available, extract listed psm.tsv files. If ALL of these are valid paths
for files within the provided FragPipe output directory (\code{path} parameter); assume this is the list
of PSM files we need. Print a warning if the table is in invalid format or if any of the listed
PSM files cannot be found within \code{path}.

As a fallback, search all direct subdirectories (i.e. not recursive) of \code{path} for psm.tsv files.
}
