% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_map_combined_protein}
\alias{fragpipe_map_combined_protein}
\title{Read FragPipe combined_protein.tsv table and then update all protein_id in provided peptide table}
\usage{
fragpipe_map_combined_protein(x, file_combined_protein)
}
\arguments{
\item{x}{peptide table composed at the end of \code{import_dataset_fragpipe_ionquant()} or \code{import_dataset_fragpipe_quantfiles()}}

\item{file_combined_protein}{see \code{import_dataset_fragpipe_ionquant()}}
}
\description{
Read FragPipe combined_protein.tsv table and then update all protein_id in provided peptide table
}
