% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe_legacy.R
\name{fragpipe_modification_to_instruction__legacy}
\alias{fragpipe_modification_to_instruction__legacy}
\title{THIS FUNCTION SHOULD ONLY BE USED FOR LEGACY FRAGPIPE DATASETS; HELPER FOR EXTRACTING PROTEIN MODIFICATIONS}
\usage{
fragpipe_modification_to_instruction__legacy(modifications)
}
\arguments{
\item{modifications}{characer vector}
}
\description{
THIS FUNCTION SHOULD ONLY BE USED FOR LEGACY FRAGPIPE DATASETS; HELPER FOR EXTRACTING PROTEIN MODIFICATIONS
}
