% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_modseq_compose}
\alias{fragpipe_modseq_compose}
\title{Reconstruct modified peptide sequences from disjoint information provided in psm.tsv to facilitate cross-reference matching of between FragPipe output tables}
\usage{
fragpipe_modseq_compose(x)
}
\arguments{
\item{x}{psm table obtained via \code{fragpipe_parse_psm()}}
}
\description{
Example data @ unchopped table in this function
HQGVMVGMGQK           5M(15.9949), 8M(15.9949) 5M(15.9949)  FALSE    FALSE    5     15.9949
HQGVMVGMGQK           5M(15.9949), 8M(15.9949) 8M(15.9949)  FALSE    FALSE    8     15.9949
ATEHPEPPK             N-term(42.0106)              N-term(42.0106)
ATTATMATSGSAR         6M(15.9949), N-term(42.0106) 6M(15.9949)
ATTATMATSGSAR         6M(15.9949), N-term(42.0106) N-term(42.0106)
}
\details{
FragPipe documentation on modified sequences (that we here replace); https://github.com/Nesvilab/MSFragger/issues/138
}
