% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_parse_msstats}
\alias{fragpipe_parse_msstats}
\title{Parse a FragPipe MSstats.csv table}
\usage{
fragpipe_parse_msstats(filename)
}
\arguments{
\item{filename}{full path to a FragPipe MSstats.csv file}
}
\description{
Parse a FragPipe MSstats.csv table
}
