% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_parse_psm}
\alias{fragpipe_parse_psm}
\title{Parse a FragPipe psm.tsv table}
\usage{
fragpipe_parse_psm(filename, precursor_unique_results = "fast_min_confidence")
}
\arguments{
\item{filename}{full path to a FragPipe psm.tsv file}

\item{precursor_unique_results}{how to deal with duplicate peptide_id*sample_id entries.
fast_min_confidence = just return min confidence (default: fast, but note that this ignores intensity values altogether)
min_confidence_max_intensity find the local minimum/maximum of respective values (pretty fast, bit slower than first option)
min_confidence_sum_intensity find the local minimum for confidence, summarize the intensities (slow)}
}
\description{
Parse a FragPipe psm.tsv table
}
