% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_parse_quant}
\alias{fragpipe_parse_quant}
\title{Parse a FragPipe quant.csv table}
\usage{
fragpipe_parse_quant(filename)
}
\arguments{
\item{filename}{full path to a FragPipe \if{html}{\out{<rawfile>}}_quant.csv file}
}
\description{
Parse a FragPipe quant.csv table
}
