% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{fragpipe_peptideresults_to_dataset}
\alias{fragpipe_peptideresults_to_dataset}
\title{From FragPipe peptide result table to an MS-DAP dataset}
\usage{
fragpipe_peptideresults_to_dataset(
  tib_result,
  acquisition_mode,
  confidence_threshold,
  collapse_peptide_by
)
}
\arguments{
\item{tib_result}{peptide table composed at the end of \code{import_dataset_fragpipe_ionquant()} or \code{import_dataset_fragpipe_quantfiles()}}

\item{acquisition_mode}{see \code{import_dataset_fragpipe_ionquant()}}

\item{confidence_threshold}{see \code{import_dataset_fragpipe_ionquant()}}

\item{collapse_peptide_by}{see \code{import_dataset_fragpipe_ionquant()}}
}
\description{
After composing a protein_id,peptide_id,sample_id,intensity,etc... table upstream, fix protein identifiers,
transform RT from seconds to minutes, transform intensity to log2, collapses peptides by plainseq/modseq and finally construct a dataset.
}
