% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_rmarkdown.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Create the Quality Control report}
\usage{
generate_pdf_report(
  dataset,
  output_dir,
  norm_algorithm = "vwmb",
  rollup_algorithm = "maxlfq",
  pca_sample_labels = "auto",
  var_explained_sample_metadata = NULL
)
}
\arguments{
\item{dataset}{your dataset}

\item{output_dir}{output directory where all output files are stored, must be an existing directory}

\item{norm_algorithm}{normalization algorithm(s). See \code{\link{analysis_quickstart}} function documentation}

\item{rollup_algorithm}{peptide-to-protein rollup strategy. See \code{\link{analysis_quickstart}} function documentation}

\item{pca_sample_labels}{configuration for PCA plot labels. See \code{\link{analysis_quickstart}} function documentation}

\item{var_explained_sample_metadata}{configuration for variance-explained analyses. See \code{\link{analysis_quickstart}} function documentation}
}
\description{
Normally used as part of the analysis_quickstart() function, refer to its implementation for example code.
}
