% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{get_column_intensity}
\alias{get_column_intensity}
\title{returns named variable variable (label=column_name) indicating which type of intensity data is used}
\usage{
get_column_intensity(peptides, contr_lbl = NA)
}
\arguments{
\item{peptides}{todo}

\item{contr_lbl}{todo}
}
\description{
returns named variable variable (label=column_name) indicating which type of intensity data is used
}
