% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{get_global_option_boolean}
\alias{get_global_option_boolean}
\title{get a value set using options(), throwing error for invalid type (boolean)}
\usage{
get_global_option_boolean(x)
}
\arguments{
\item{x}{option name for \code{getOption()}}
}
\description{
get a value set using options(), throwing error for invalid type (boolean)
}
