% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{get_mode}
\alias{get_mode}
\title{find the mode in a numeric array using the density function from base R}
\usage{
get_mode(x)
}
\arguments{
\item{x}{numeric array}
}
\value{
mode of x
}
\description{
find the mode in a numeric array using the density function from base R
}
