% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{get_peptide_filternorm_variants}
\alias{get_peptide_filternorm_variants}
\title{Return column names in peptide table that hold intensity values}
\usage{
get_peptide_filternorm_variants(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
Return column names in peptide table that hold intensity values
}
