% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_protein_matrix}
\alias{get_protein_matrix}
\title{rollup peptides to protein data matrix for selected intensity column}
\usage{
get_protein_matrix(dataset, intensity_column, include_npep = TRUE)
}
\arguments{
\item{dataset}{your dataset}

\item{intensity_column}{column in \code{dataset$peptides} that should be used for the protein matrix}

\item{include_npep}{if \code{TRUE} (default), returns a list with both the matrix and an array of peptide counts (for each row in the protein matrix). If \code{FALSE}, returns just the protein matrix}
}
\description{
rollup peptides to protein data matrix for selected intensity column
}
