% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_samples_for_regression}
\alias{get_samples_for_regression}
\title{collect the respective subset of samples for selected protein data and prepare a metadata table for regression}
\usage{
get_samples_for_regression(dataset, protein_data)
}
\arguments{
\item{dataset}{your dataset}

\item{protein_data}{output from \code{get_protein_matrix()}}
}
\description{
From sample metadata (dataset$samples), get the subset that is included in the
parameter \code{protein_data}, select columns that contain potential regression variables
(using \code{user_provided_metadata()}) and finally apply \code{enforce_sample_value_types()} to
reformat all variables (except "sample_id") to factor or numeric.
}
