% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{generate colours analogous to ggplot's default palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
https://stackoverflow.com/a/8197703
}
