% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{ggplot_peptide}
\alias{ggplot_peptide}
\title{placeholder title}
\usage{
ggplot_peptide(
  protein_id,
  filename,
  plot_title,
  tib_peptides,
  tib_stats,
  tib_sample_groups
)
}
\arguments{
\item{protein_id}{todo}

\item{filename}{todo}

\item{plot_title}{todo}

\item{tib_peptides}{todo}

\item{tib_stats}{todo}

\item{tib_sample_groups}{todo}
}
\description{
placeholder title
}
