% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{ggplot_peptide_abundances}
\alias{ggplot_peptide_abundances}
\title{plot peptide-level data for a single protein}
\usage{
ggplot_peptide_abundances(
  tib_peptides,
  tib_samples,
  tib_stats,
  plot_title = ""
)
}
\arguments{
\item{tib_peptides}{specifically crafted peptides tibble, see calling functions for specs}

\item{tib_samples}{table with all samples for current plot}

\item{tib_stats}{table with statistics shown below plot}

\item{plot_title}{optionally a plot title}
}
\description{
plot peptide-level data for a single protein
}
