% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_metadata.R
\name{ggplot_sample_detect_counts_barplots}
\alias{ggplot_sample_detect_counts_barplots}
\title{placeholder title}
\usage{
ggplot_sample_detect_counts_barplots(samples, samples_colors_long)
}
\arguments{
\item{samples}{todo}

\item{samples_colors_long}{todo}
}
\description{
placeholder title
}
