% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_metadata.R
\name{ggplot_sample_detect_vs_metadata_scatterplot}
\alias{ggplot_sample_detect_vs_metadata_scatterplot}
\title{compact scatterplot with color-coding captures all sample metadata in one figure}
\usage{
ggplot_sample_detect_vs_metadata_scatterplot(tib_plot)
}
\arguments{
\item{tib_plot}{todo}
}
\description{
compact scatterplot with color-coding captures all sample metadata in one figure
}
