% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_metadata.R
\name{ggplot_sample_detect_vs_metadata_scatterplot_by_prop}
\alias{ggplot_sample_detect_vs_metadata_scatterplot_by_prop}
\title{separate plot for each metadata property}
\usage{
ggplot_sample_detect_vs_metadata_scatterplot_by_prop(tib_plot)
}
\arguments{
\item{tib_plot}{todo}
}
\description{
separate plot for each metadata property
}
