% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rmarkdown.R
\name{ggplot_split_legend}
\alias{ggplot_split_legend}
\title{split a ggplot object into 2 ggplots; original plot minus legend and only the legend}
\usage{
ggplot_split_legend(p)
}
\arguments{
\item{p}{ggplot object}
}
\description{
split a ggplot object into 2 ggplots; original plot minus legend and only the legend
}
