% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{gui}
\alias{gui}
\title{open the MS-DAP graphical user interface (a R shiny app)}
\usage{
gui()
}
\arguments{
\item{...}{optionally, pass some additional parameters to shiny::shinyApp()}
}
\description{
open the MS-DAP graphical user interface (a R shiny app)
}
