% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{has_legacy_contrast_definitions}
\alias{has_legacy_contrast_definitions}
\title{test if the dataset has contrast definitions that are incompatible with MS-DAP version 1.2}
\usage{
has_legacy_contrast_definitions(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
test if the dataset has contrast definitions that are incompatible with MS-DAP version 1.2
}
