% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{hgnc_add_xrefs}
\alias{hgnc_add_xrefs}
\title{add hgnc_symbol and entrez/ensembl IDs if present in hgnc table}
\usage{
hgnc_add_xrefs(x, hgnc)
}
\arguments{
\item{x}{a data.table with column hgnc_id}

\item{hgnc}{HGNC lookup table from \code{hgnc_lookuptable()}}
}
\description{
add hgnc_symbol and entrez/ensembl IDs if present in hgnc table
}
