% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{idmap_symbols}
\alias{idmap_symbols}
\title{helper function to match a protein table to HGNC via gene symbols (and synonyms)}
\usage{
idmap_symbols(proteins_long, hgnc, use_synonyms = TRUE)
}
\arguments{
\item{proteins_long}{long-format data.frame with columns protein_id and symbol pairs (i.e. no semicolon-delimited values)}

\item{hgnc}{HGNC lookup table from \code{hgnc_lookuptable()}}

\item{use_synonyms}{boolean; use synonyms? if FALSE, only uses hgnc_symbol}
}
\description{
typically called from \code{export_stats_genesummary}
}
