% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{idmap_uniprotid}
\alias{idmap_uniprotid}
\title{helper function to match a protein table to HGNC via an intermediate mapping table (e.g. MGI/RGD)}
\usage{
idmap_uniprotid(proteins_long, hgnc, idmap, use_symbols = TRUE)
}
\arguments{
\item{proteins_long}{long-format data.frame with columns uniprot_id and symbol pairs (i.e. no semicolon-delimited values)}

\item{hgnc}{HGNC lookup table from \code{hgnc_lookuptable()}}

\item{idmap}{MGI or RGD lookup table from \code{mgi_lookuptable()} or \code{rgd_lookuptable()}}

\item{use_symbols}{boolean; use symbols to map from proteins_long to the idmap table? if FALSE, only matches by uniprot identifier}
}
\description{
typically called from \code{export_stats_genesummary}
}
