% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dataset_wrapper.R
\name{import_dataset__generic}
\alias{import_dataset__generic}
\title{A generic wrapper around data import functions included with MS-DAP}
\usage{
import_dataset__generic(path, type)
}
\arguments{
\item{path}{the full file path of the input file or directory}

\item{type}{the type of dataset you want to import}
}
\description{
This function simply calls the respective implementation with default parameters. In most cases it's more appropriate to explicitly call the respective import function for your data, e.g. import_dataset_skyline(). The main use-case of this function is to convenience implementation in the Graphical User Interface (GUI).
}
