% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longformat_generic.R
\name{import_dataset_diann}
\alias{import_dataset_diann}
\title{Import a label-free proteomics dataset from DIA-NN}
\usage{
import_dataset_diann(
  filename,
  confidence_threshold = 0.01,
  use_normalized_intensities = FALSE,
  use_irt = TRUE,
  select_unique_precursor_per_modseq = TRUE,
  do_plot = TRUE
)
}
\arguments{
\item{filename}{the full file path of the DIA-NN report; this can be either the report.tsv file or the report.parquet file (new since DIA-NN 1.9.1)}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified', default: 0.01 (target value must be lesser than or equals)}

\item{use_normalized_intensities}{use the abundance values as-is (recommended) or those normalized by DIA-NN}

\item{use_irt}{logical indicating whether to use standardized (IRT) or the empirical (RT) retention times}

\item{select_unique_precursor_per_modseq}{if multiple precursors are available for a modified sequence (eg; charge 2 and 3), should we make an effort to select the best or just return all of them ?}

\item{do_plot}{logical indicating whether to create QC plots that are shown in the downstream PDF report (if enabled)}
}
\description{
Import a label-free proteomics dataset from DIA-NN
}
