% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_encyclopedia.R
\name{import_dataset_encyclopedia}
\alias{import_dataset_encyclopedia}
\title{Import a label-free proteomics dataset from EncyclopeDIA}
\usage{
import_dataset_encyclopedia(
  path,
  confidence_threshold = 0.01,
  return_decoys = FALSE
)
}
\arguments{
\item{path}{the directory that contains the EncyclopeDIA ELIB files}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{return_decoys}{logical indicating whether to return decoy peptides. Should be set to FALSE, and if enabled, make sure to manually remove the decoys from the peptides tibble before running the quickstart function!}
}
\description{
This function combines all 'ELIB' files from an EncyclopeDIA analysis into a dataset.
}
\details{
Input files in the provided input directory should be the .elib files matching each sample (mzML/dia input file to EncyclopeDIA)
}
