% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{import_dataset_fragpipe_psm_file}
\alias{import_dataset_fragpipe_psm_file}
\title{Alternative FragPipe workflow, operates directly on psm.tsv file}
\usage{
import_dataset_fragpipe_psm_file(
  filename,
  acquisition_mode,
  confidence_threshold = 0.01,
  collapse_peptide_by = "sequence_modified",
  return_decoys = FALSE,
  do_plot = TRUE
)
}
\arguments{
\item{filename}{the full file path of the input file}

\item{acquisition_mode}{the type of experiment, should be a string. options: "dda" or "dia"}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}

\item{return_decoys}{logical indicating whether to return decoy peptides. Should be set to FALSE, and if enabled, make sure to manually remove the decoys from the peptides tibble before running the quickstart function!}

\item{do_plot}{logical indicating whether to create QC plots that are shown in the downstream PDF report (if enabled)}
}
\description{
For typical FragPipe workflows, import_dataset_fragpipe_ionquant() should be used
instead of this function.
}
\details{
This function directly parses data from the psm.tsv and doesn't use MBR data
presented in other files. The most abundant PSM per peptide*sample is picked to
represent abundance.
}
