% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longformat_generic.R
\name{import_dataset_in_long_format}
\alias{import_dataset_in_long_format}
\title{For advanced users: A generic function for importing data in long-format, typically used for DIA data.}
\usage{
import_dataset_in_long_format(
  filename = NULL,
  x = NULL,
  attributes_required = list(),
  attributes_optional = list(),
  confidence_threshold = 0.01,
  remove_peptides_never_above_confidence_threshold = TRUE,
  select_unique_precursor_per_modseq = TRUE,
  custom_func_manipulate_DT_onload = NULL,
  return_decoys = TRUE,
  do_plot = TRUE
)
}
\arguments{
\item{filename}{the full file path of the input file. Either supply a filename, or a data table}

\item{x}{input data table as a data.frame (or tibble, or data.table). Either supply a filename, or a data table}

\item{attributes_required}{a list of attributes that must be present in the input file (eg; sample names, peptide IDs, intensity values, etc.)}

\item{attributes_optional}{a list of optional attributes that we can do without downstream, like m/z}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{remove_peptides_never_above_confidence_threshold}{set to true to remove all target peptides that never pass the confidence score threshold}

\item{select_unique_precursor_per_modseq}{if multiple precursors are available for a modified sequence (eg; charge 2 and 3), should we make an effort to select the best or just return all of them ?}

\item{custom_func_manipulate_DT_onload}{!expert use only! (there is intricate interplay between such pre-filtering and the code in this function); provide a function that manipulates the data.table after load. example implementation in the fragpipe parser.}

\item{return_decoys}{logical indicating whether to return decoy peptides. Should be set to FALSE, and if enabled, make sure to manually remove the decoys from the peptides tibble before running the quickstart function!}

\item{do_plot}{logical indicating whether to create QC plots that are shown in the downstream PDF report (if enabled)}
}
\description{
Refer to the implementation of \code{import_dataset_spectronaut} for an example on how to use this function.
don't forget to set either "dia" or "dda" for dataset$acquisition_mode downstream if you implement this.
}
