% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_maxquant.R
\name{import_dataset_maxquant_evidencetxt}
\alias{import_dataset_maxquant_evidencetxt}
\title{Import a label-free proteomics dataset from MaxQuant}
\usage{
import_dataset_maxquant_evidencetxt(
  path,
  collapse_peptide_by = "sequence_modified",
  remove_shared = TRUE
)
}
\arguments{
\item{path}{the directory that contains the search results (typically the 'txt' filter that contains files 'proteinGroups.txt' and 'evidence.txt')}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}

\item{remove_shared}{remove all shared peptides that could not be assigned to a single proteingroup as a unique- or razor-peptide ?}
}
\description{
Import a label-free proteomics dataset from MaxQuant
}
