% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metamorpheus.R
\name{import_dataset_metamorpheus}
\alias{import_dataset_metamorpheus}
\title{Import a label-free proteomics dataset from MetaMorpheus}
\usage{
import_dataset_metamorpheus(
  path,
  protein_qval_threshold = 0.05,
  collapse_peptide_by = "sequence_modified",
  only_unique_peptides = FALSE
)
}
\arguments{
\item{path}{the directory that contains the search results (eg; present files are AllProteinGroups.tsv, AllQuantifiedPeaks.tsv, etc.)}

\item{protein_qval_threshold}{qvalue threshold for accepting target proteins}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}

\item{only_unique_peptides}{remove all shared peptides? recommended to leave at FALSE}
}
\description{
Import a label-free proteomics dataset from MetaMorpheus
}
