% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openms.R
\name{import_dataset_openms_mztab}
\alias{import_dataset_openms_mztab}
\title{Import a label-free DDA proteomics dataset from OpenMS, experimental feature !}
\usage{
import_dataset_openms_mztab(filename)
}
\arguments{
\item{filename}{full path to the mzTab file}
}
\description{
This parser is based on output we generated with OpenMS 2.5 using input data and parameters/instructions from: https://abibuilder.informatik.uni-tuebingen.de/archive/openms/Tutorials/Data/iPRG2015_Hannes/
}
\details{
Check the file "run_lfq.sh" in the above instructions (from OpenMS team) for exact workflow and settings used.

By lack of more example mzTab files that we can adapt to, this code now \strong{only} supports input data generated as in run_lfq.sh from the above example;

The main dependency / complication is that we require the peptide- and protein-scores exactly as computed in the reference OpenMS workflow.
Thus, on each mzML file, apply search engine (eg; xtandem), PercolatorAdapter, FalseDiscoveryRate, IDFilter, IDScoreSwitcher.
finally run ProteomicsLFQ on the entire dataset.
}
