% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_peaks.R
\name{import_dataset_peaks}
\alias{import_dataset_peaks}
\title{Import a label-free proteomics dataset from Peaks}
\usage{
import_dataset_peaks(filename, collapse_peptide_by = "sequence_modified")
}
\arguments{
\item{filename}{a features.csv file exported by Peaks}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}
}
\description{
Import a label-free proteomics dataset from Peaks
}
