% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_proteomediscoverer_txt.R
\name{import_dataset_proteomediscoverer_txt}
\alias{import_dataset_proteomediscoverer_txt}
\title{Import a label-free DDA proteomics dataset from ProteomeDiscoverer, experimental feature !}
\usage{
import_dataset_proteomediscoverer_txt(filename)
}
\arguments{
\item{filename}{full path to the _PSMs.txt file}
}
\description{
Input data must contain Percolator PEP scores for each PSM, so after search engine (eg; Sequest HT) make sure to connect the Percolator node.
}
\details{
Example PD workflow:
Processing Step: PWF_QE_Precursor_Quan_and_LFQ_SequestHT_Percolator
Consensus Step: CWF_Comprehensive_Enhanced Annotation_LFQ_and_Precursor_Quan
Consensus Step: add the "result exporter" (drag&drop from side panel to bottom panel)

Optionally, you can relax the output filter criteria somewhat, thereby relying on MS-DAP filtering a bit more and potentially gaining some extra (MBR)hits, as follows;
Consensus step -->> "peptide and protein filter" -->> Peptide Confidence At Least -->> change to medium
While you're making changes there, you can problably set "Remove Peptides Without Proteins" to "True"
}
