% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longformat_generic.R
\name{import_dataset_proteoscape}
\alias{import_dataset_proteoscape}
\title{Import a label-free proteomics dataset from Bruker proteoscape}
\usage{
import_dataset_proteoscape(
  filename_peptides,
  filename_proteins,
  use_normalized_intensities = FALSE,
  confidence_threshold = 0.01,
  use_irt = TRUE,
  select_unique_precursor_per_modseq = TRUE,
  do_plot = TRUE
)
}
\arguments{
\item{filename_peptides}{the full file path of the peptides parquet file}

\item{filename_proteins}{the full file path of the proteins parquet file}

\item{use_normalized_intensities}{use the abundance values as-is (recommended) or those normalized by DIA-NN}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified', default: 0.01 (target value must be lesser than or equals)}

\item{use_irt}{logical indicating whether to use standardized (IRT) or the empirical (RT) retention times}

\item{select_unique_precursor_per_modseq}{if multiple precursors are available for a modified sequence (eg; charge 2 and 3), should we make an effort to select the best or just return all of them ?}

\item{do_plot}{logical indicating whether to create QC plots that are shown in the downstream PDF report (if enabled)}
}
\description{
This function was tested with a Proteoscape that uses DIA-NN 1.7.1, and expects the
respective peptide.parquet and protein.parquet files with peptide and protein data.
}
\details{
You can also import the report.tsv file generated by Proteoscape using the \code{import_dataset_diann}
function, which is a lot faster (because we don't have to post-process the Proteoscape-specific output files).
}
