% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expressionset.R
\name{import_expressionset}
\alias{import_expressionset}
\title{Import a label-free proteomics dataset from a peptide-level Biobase ExpressionSet object}
\usage{
import_expressionset(
  eset,
  column_fdata_protein_id = "prot.id",
  column_pdata_sample_group = "condition",
  acquisition_mode = "dda",
  is_log2 = FALSE
)
}
\arguments{
\item{eset}{a Biobase ExpressionSet containing peptide data}

\item{column_fdata_protein_id}{the column name in Biobase::fData() that holds the protein_id}

\item{column_pdata_sample_group}{the column name in Biobase::pData() that holds the sample group/condition}

\item{acquisition_mode}{the type of label-free data. options: 'dda' or 'dia'}

\item{is_log2}{logical indicating whether the provided intensities are already log2 transformed}
}
\description{
provided mostly for compatability with results from prior bioinformatic analyses, as it is preferred to import raw data files.
}
