% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{import_fasta}
\alias{import_fasta}
\title{Import fasta file(s) that match your dataset.}
\usage{
import_fasta(
  dataset,
  files = NULL,
  fasta_id_type = "uniprot",
  protein_separation_character = ";"
)
}
\arguments{
\item{dataset}{your dataset}

\item{files}{an array of filenames, these should be the full path}

\item{fasta_id_type}{what type of fasta files are these? options: "uniprot" (highly recommended) or otherwise any other character string (as we have no special rules for generic fasta files)}

\item{protein_separation_character}{the separation character for protein identifiers in your dataset. Most commonly this is a semicolon (eg; in maxquant/metamorpheus/skyline/etc.)}
}
\description{
For fasta files from uniprot, we extract the gene symbols for each protein-group (not available for non-uniprot fasta files).
}
