% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_maxquant.R
\name{import_maxquant_peptides}
\alias{import_maxquant_peptides}
\title{Import a label-free proteomics dataset from MaxQuant, using the peptides.txt file}
\usage{
import_maxquant_peptides(
  file_peptides,
  remove_shared_peptides = T,
  pep2prot = NA
)
}
\arguments{
\item{file_peptides}{full file path to MaxQuant's peptides.txt output file}

\item{remove_shared_peptides}{remove all peptides that are assigned to multiple proteingroups (checked by having a semicolon in the 'Protein group IDs' column)}

\item{pep2prot}{if available, provide a 2 column table that provides a mapping from each peptide to a protein. Required columns; protein_id and sequence_plain. Set NA to use "Leading razor protein" as protein grouping}
}
\description{
Recommended to import directly from the evidence file using \code{import_dataset_maxquant_evidencetxt} instead, as the peptides.txt file lacks retention-tine information !
}
\details{
Provided mostly for compatability with incomplete datasets that lack the evidence.txt and proteinGroups.txt files
}
