% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_maxquant.R
\name{import_maxquant_proteingroups}
\alias{import_maxquant_proteingroups}
\title{Import peptide-to-protein mappings from MaxQuant}
\usage{
import_maxquant_proteingroups(path, remove_shared = T)
}
\arguments{
\item{path}{the directory that contains the search results (typically the 'txt' filter that contains files 'proteinGroups.txt' and 'evidence.txt')}

\item{remove_shared}{remove all shared peptides that could not be assigned to a single proteingroup as a unique- or razor-peptide ?}
}
\description{
Import peptide-to-protein mappings from MaxQuant
}
