% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metamorpheus.R
\name{import_metamorpheus_proteingroups}
\alias{import_metamorpheus_proteingroups}
\title{placeholder title}
\usage{
import_metamorpheus_proteingroups(
  file_prot,
  protein_qval_threshold = 0.05,
  only_unique_peptides = FALSE
)
}
\arguments{
\item{file_prot}{todo}

\item{protein_qval_threshold}{todo}

\item{only_unique_peptides}{todo}
}
\description{
placeholder title
}
