% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metamorpheus.R
\name{import_metamorpheus_quantifiedpeaks}
\alias{import_metamorpheus_quantifiedpeaks}
\title{placeholder title}
\usage{
import_metamorpheus_quantifiedpeaks(file_quantpeaks)
}
\arguments{
\item{file_quantpeaks}{todo}
}
\description{
placeholder title
}
