% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{import_protein_metadata_from_fasta}
\alias{import_protein_metadata_from_fasta}
\title{placeholder title}
\usage{
import_protein_metadata_from_fasta(
  protein_id,
  fasta_files,
  fasta_id_type = "uniprot",
  protein_separation_character = ";"
)
}
\arguments{
\item{protein_id}{todo}

\item{fasta_files}{todo}

\item{fasta_id_type}{todo}

\item{protein_separation_character}{todo}
}
\value{
table where
protein_id = provided proteingroup identifier,
accessions = result from fasta_id_short applied to each semicolon-delimited element in protein_id (result is a semicolon-collapsed string),
fasta_headers = analogous to accessions, but matching the full FASTA header to each 'accessions' element,
gene_symbols = full set of gene symbols, '-' where missing in FASTA, matching each element in 'accessions',
gene_symbols_or_id = unique set of valid 'gene_symbol', or the FASTA full/long ID when there is no gene information
gene_symbol_ucount = number of unique gene_symbols for this proteingroup (i.e. unique valid elements in 'gene_symbols')
}
\description{
placeholder title
}
