% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{import_sample_metadata}
\alias{import_sample_metadata}
\title{Import sample metadata from an Excel table}
\usage{
import_sample_metadata(dataset, filename)
}
\arguments{
\item{dataset}{your dataset}

\item{filename}{full file path of the input file (eg; C:/temp/template_experiment1.xlsx)}
}
\description{
Import sample metadata from an Excel table
}
