% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{initialize_multiprocessing}
\alias{initialize_multiprocessing}
\title{To facilitate multiprocessing, create a Parallel Socket Cluster object}
\usage{
initialize_multiprocessing(n_thread = 0)
}
\arguments{
\item{n_thread}{maximum number of threads}
}
\description{
You only need this if a) not using the quickstart function and b) running DEA algorithms msqrob or msqrobsum
}
