% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{is_dia_dataset}
\alias{is_dia_dataset}
\title{Test if the dataset contains DIA data, simply testing if the acquisition_mode equals "dia" (case insensitive)}
\usage{
is_dia_dataset(dataset)
}
\arguments{
\item{dataset}{dataset to test}
}
\description{
Test if the dataset contains DIA data, simply testing if the acquisition_mode equals "dia" (case insensitive)
}
