% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{limma_fit_extract_stats}
\alias{limma_fit_extract_stats}
\title{extract limma fit results}
\usage{
limma_fit_extract_stats(fit, contrast_name = NULL)
}
\arguments{
\item{fit}{result from \code{limma::lmFit()}}

\item{contrast_name}{name of a specific contrast(s) to extract. Default value, \code{NULL}, will return results from all contrasts}
}
\description{
extract limma fit results
}
