% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{limma_wrapper}
\alias{limma_wrapper}
\title{Perform linear regression with limma}
\usage{
limma_wrapper(
  x = protein_data,
  npep = NULL,
  model_matrix,
  contrasts,
  limma_block_variable = NULL,
  ebayes_trend = FALSE,
  ebayes_robust = FALSE,
  deqms = TRUE,
  return_table = TRUE
)
}
\arguments{
\item{x}{result from \code{get_protein_matrix(... , include_npep = TRUE)}. Alternatively, a protein log2 intensity matrix (in this case, also provide the \code{npep} parameter if you want to apply DEqMS)}

\item{npep}{optionally, the number of peptides-per-protein for each row in input protein matrix. Only needed when providing a matrix for parameter \code{x} (i.e. you can leave this NULL when \code{x} is output from \code{get_protein_matrix(... , include_npep = TRUE)})}

\item{model_matrix}{result from \code{stats::model.matrix()}}

\item{contrasts}{contrasts defined with \code{limma::makeContrasts()} that should be extracted from the linear regression model}

\item{limma_block_variable}{option passed to \code{limma::lmFit()} to fit block design. To skip (default), set NULL. Otherwise, this should be a character vector of the same length as nrow(model_matrix) so it describes a block/group per respective sample in the model/design matrix}

\item{ebayes_trend}{option passed to \code{limma::eBayes()}. Defaults to \code{FALSE} (same as limma default)}

\item{ebayes_robust}{option passed to \code{limma::eBayes()}. Defaults to \code{FALSE} (same as limma default)}

\item{deqms}{apply DEqMS? Defaults to \code{TRUE}. If \code{FALSE}, returns \code{limma::eBayes()} result as-is}

\item{return_table}{if \code{TRUE}, the default, returns a table with protein_id, log2fc, pvalue, adjusted pvalue, contrast (i.e. this is the output from MS-DAP function \code{limma_fit_extract_stats()}). If \code{FALSE}, returns the limma fit object as-is}
}
\description{
Please refer to the vignette on custom limma models at the MS-DAP GitHub repository for a walk-through with examples.
}
