% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{log2_to_ln}
\alias{log2_to_ln}
\title{convert base-log from log2 to natural log. log_a(x) = log_b(x) / log_b(a)
example: log2(4); exp(log2_to_ln(log2(4)))}
\usage{
log2_to_ln(x)
}
\arguments{
\item{x}{numeric values that were log2 transformed}
}
\description{
convert base-log from log2 to natural log. log_a(x) = log_b(x) / log_b(a)
example: log2(4); exp(log2_to_ln(log2(4)))
}
