% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{log_peptide_tibble_pep_prot_counts}
\alias{log_peptide_tibble_pep_prot_counts}
\title{report number of detected precursor and plainseq, target and decoy}
\usage{
log_peptide_tibble_pep_prot_counts(peptides)
}
\arguments{
\item{peptides}{peptide tibble in long format}
}
\description{
report number of detected precursor and plainseq, target and decoy
}
