% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_log.R
\name{logger.default}
\alias{logger.default}
\title{Log function that takes any string, appends it to global log (log_ , a list) and prints to console}
\usage{
logger.default(x, type = "info")
}
\arguments{
\item{x}{your message as a character string}

\item{type}{the type of message. options: info, progress, warning, error. For the latter, an R error is thrown after logging!}
}
\description{
Log function that takes any string, appends it to global log (log_ , a list) and prints to console
}
