% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{make_bins}
\alias{make_bins}
\title{bin an array and return both the bins and their metadata (label and mean value)}
\usage{
make_bins(x, from, to, length)
}
\arguments{
\item{x}{numeric vector}

\item{from}{parameter for seq()}

\item{to}{parameter for seq()}

\item{length}{length.out parameter for seq()}
}
\description{
bin an array and return both the bins and their metadata (label and mean value)
}
