% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{match_list_elements_to_vector}
\alias{match_list_elements_to_vector}
\title{placeholder title}
\usage{
match_list_elements_to_vector(list.of.strings, referenceList, remove.na = T)
}
\arguments{
\item{list.of.strings}{todo}

\item{referenceList}{todo}

\item{remove.na}{todo}
}
\description{
placeholder title
}
