% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{matrix_grouped_column_aggregate}
\alias{matrix_grouped_column_aggregate}
\title{fast grouped matrix aggregation using data.table}
\usage{
matrix_grouped_column_aggregate(x, x_rowgroups, FUN)
}
\arguments{
\item{x}{a matrix}

\item{x_rowgroups}{grouping variable}

\item{FUN}{function to apply, e.g. \code{sum}}
}
\value{
a matrix, where rownames match the (unique values of the) grouping variable
}
\description{
fast grouped matrix aggregation using data.table
}
\examples{
\dontrun{
x = my_integer_matrix
x_rowgroups = my_grouping_var
result = matrix_grouped_column_aggregate(x, x_rowgroups, FUN = sum)

# validate this function yields the same as (slower) stats::aggregate()
df = stats::aggregate(x, by = list(groupid = x_rowgroups), FUN = sum)
i = match(rownames(result), df$groupid)
stopifnot(!is.na(i))
stopifnot(count_detect == as.matrix(df[,-1])[i,])
stopifnot(rownames(count_detect) == df$groupid[i])
}
}
