% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{matrix_to_long}
\alias{matrix_to_long}
\title{placeholder title}
\usage{
matrix_to_long(
  mat,
  value_name = "value",
  column_name = "sample",
  row_name = "sequence"
)
}
\arguments{
\item{mat}{todo}

\item{value_name}{todo}

\item{column_name}{todo}

\item{row_name}{todo}
}
\description{
placeholder title
}
