% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{merge_fractionated_samples}
\alias{merge_fractionated_samples}
\title{merge fractionated samples}
\usage{
merge_fractionated_samples(dataset)
}
\arguments{
\item{dataset}{a valid dataset}
}
\description{
to collapse sample fractions, the sample metadata table must contain columns 'fraction' and 'shortname' (eg; sample_id is unique sample/measurement, whereas shortname is the identifier to which respective fractions are collapsed to)
}
