% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{metadata_matrix_from_filenames}
\alias{metadata_matrix_from_filenames}
\title{Guess experiment data from a set of filenames}
\usage{
metadata_matrix_from_filenames(x)
}
\description{
Assuming input files are somewhat well formatted,
we here try to detect sample groups and other metadata to populate metadata table for user's convenience
}
\details{
example; wt1, wt2, ko1, ko2, ...
example; wt_1, wt_2, ko_1, ko_2, ...
example; control_day1_1, control_day1_2, cond1_day1_1, cond1_day2_2, cond1_day2_3, ...
}
