% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{mgi_lookuptable}
\alias{mgi_lookuptable}
\title{Parse MGI ID mapping table table into a lookup table with symbols and uniprot IDs}
\usage{
mgi_lookuptable(f)
}
\arguments{
\item{f}{full path to the downloaded table (expected to be tsv format)}
}
\description{
download link: https://www.informatics.jax.org/downloads/reports/index.html
table: "MGI Marker associations to SWISS-PROT and TrEMBL protein IDs (tab-delimited)"
filename is typically something like MRK_SwissProt_TrEMBL.rpt
}
