% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{minlog10}
\alias{minlog10}
\title{Accept pvalues that are exactly zero as the 'best pvalue'}
\usage{
minlog10(x)
}
\arguments{
\item{x}{values to transform}
}
\description{
Accept pvalues that are exactly zero as the 'best pvalue'
}
