% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdap-package.R
\docType{package}
\name{msdap-package}
\alias{msdap}
\alias{msdap-package}
\title{MS-DAP: Mass Spectrometry Downstream Analysis Pipeline for label-free proteomics data.}
\description{
http://github.com/ftwkoopmans/msdap
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftwkoopmans/msdap}
}

}
\author{
\strong{Maintainer}: Frank Koopmans \email{frank.koopmans@vu.nl} (\href{https://orcid.org/0000-0002-4973-5732}{ORCID})

}
\keyword{internal}
