% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{msdap_shiny_server}
\alias{msdap_shiny_server}
\title{MS-DAP shiny server}
\usage{
msdap_shiny_server(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
MS-DAP shiny server
}
