% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{norm_scales_fcmode}
\alias{norm_scales_fcmode}
\title{minimize the foldchange-mode between all columns}
\usage{
norm_scales_fcmode(x)
}
\arguments{
\item{x}{log-transformed numeric matrix}
}
\value{
scaling factor for each column in x
}
\description{
minimize the foldchange-mode between all columns
}
