% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{norm_scales_var}
\alias{norm_scales_var}
\title{minimize variation over all rows in a numeric matrix}
\usage{
norm_scales_var(x)
}
\arguments{
\item{x}{log-transformed numeric matrix}
}
\value{
scaling factor for each column in x
}
\description{
minimize variation over all rows in a numeric matrix
}
