% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalization_algorithms}
\alias{normalization_algorithms}
\title{the set of available normalization algorithms}
\usage{
normalization_algorithms()
}
\description{
the set of available normalization algorithms
}
