% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalize_median}
\alias{normalize_median}
\title{scale each sample such that median abundance values are the same for all samples in the dataset}
\usage{
normalize_median(x_as_log2, ...)
}
\arguments{
\item{x_as_log2}{log2 transformed abundance values}

\item{...}{remaining parameters are ignored}
}
\description{
scale each sample such that median abundance values are the same for all samples in the dataset
}
