% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{normalize_modebetween_protein}
\alias{normalize_modebetween_protein}
\title{Given some peptide-level matrix, normalize samples using foldchange mode-between at protein-level}
\usage{
normalize_modebetween_protein(x, proteins, groups, rollup_algorithm = "maxlfq")
}
\arguments{
\item{x}{numerical data matrix with log2 peptide abundances}

\item{proteins}{array describing the grouping of the rows in x}

\item{groups}{array describing the grouping of the columns in x}

\item{rollup_algorithm}{algorithm for combining peptides to protein-level data, on which in subsequent steps the normalization/scaling factors will be computed. Refer to \code{\link{rollup_pep2prot}} function documentation for available options and a brief description of each}
}
\value{
normalized matrix x
}
\description{
Given some peptide-level matrix, normalize samples using foldchange mode-between at protein-level
}
