% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalize_msempire}
\alias{normalize_msempire}
\title{wrapper function for msEmpiRe::normalize()}
\usage{
normalize_msempire(x, group_by_cols)
}
\arguments{
\item{x}{numerical matrix, NOT log transformed}

\item{group_by_cols}{a character array of the same length as the number of columns in the provided data matrix, describing the grouping of each column/sample}
}
\description{
wrapper function for msEmpiRe::normalize()
}
