% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{normalize_peptide_intensity_column}
\alias{normalize_peptide_intensity_column}
\title{Apply an array of normalization algorithms to some intensity column in the peptide table}
\usage{
normalize_peptide_intensity_column(
  dataset,
  col_intensity,
  norm_algorithm,
  rollup_algorithm
)
}
\arguments{
\item{dataset}{your dataset}

\item{col_intensity}{intensity column in dataset$peptides tibble}

\item{norm_algorithm}{array of normalization algorithm character IDs}

\item{rollup_algorithm}{the algorithm for combining peptides to proteins as used in normalization algorithms that require a priori rollup from peptides to a protein-level abundance matrix (e.g. modebetween_protein). Refer to \code{\link{rollup_pep2prot}} function documentation for available options and a brief description of each}
}
\description{
Apply an array of normalization algorithms to some intensity column in the peptide table
}
