% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalize_rlr}
\alias{normalize_rlr}
\title{Wrapper function for normalize_rlr_MSqRob_implementation}
\usage{
normalize_rlr(x)
}
\arguments{
\item{x}{log2 abundance matrix}
}
\description{
Wrapper function for normalize_rlr_MSqRob_implementation
}
