% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vwmb.R
\name{normalize_vwmb}
\alias{normalize_vwmb}
\title{VWMB: normalize a numerical matrix by Variation Within and Mode Between}
\usage{
normalize_vwmb(
  x,
  groups = NA,
  metric_within = "var",
  metric_between = "mode",
  include_attributes = FALSE
)
}
\arguments{
\item{x}{numerical data matrix to normalize, should be transformed by logarithm}

\item{groups}{array describing the grouping of the columns in x. set NA for no groups}

\item{metric_within}{how should replicate samples within a group be normalized? valid arguments: "var" reduce overall variation (default). "mode" reduce overall foldchange mode. pass empty string to disable}

\item{metric_between}{analogous to the metric_within parameter, how to normalize between groups? allowed parameters are "var" and "mode" (default)}

\item{include_attributes}{optionally, return some additional metrics as attributes of x. The "scaling" attribute describes the increase/decrease of each sample}
}
\value{
normalized matrix x
}
\description{
note; if you want to treat replicate samples that are flagged as 'exclude' upstream differently, set groups=paste(samples$group, samples$exclude) to put them in separate groups
}
