% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{order_gene_symbols}
\alias{order_gene_symbols}
\title{sort; mixedorder the gene symbols, but put 'has no gene symbol' (protein long ID is used as placeholder) on top}
\usage{
order_gene_symbols(s)
}
\arguments{
\item{s}{todo}
}
\description{
sort; mixedorder the gene symbols, but put 'has no gene symbol' (protein long ID is used as placeholder) on top
}
