% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{parameter_documentation}
\alias{parameter_documentation}
\title{extract documentation for all parameters of some function}
\usage{
parameter_documentation(func = NULL, package = "msdap")
}
\arguments{
\item{func}{function name as a string}

\item{package}{package name as a string}
}
\description{
returns a named array where values are help text and names are the function parameters
}
