% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{parse_fragpipe_proteins}
\alias{parse_fragpipe_proteins}
\title{parse a proteins.tsv or combined_protein.tsv from fragpipe output and return
a well formatted protein_id column that includes both "leading protein ID"
and "Indistinguishable Proteins"}
\usage{
parse_fragpipe_proteins(file_proteins)
}
\arguments{
\item{file_proteins}{full file path}
}
\description{
parse a proteins.tsv or combined_protein.tsv from fragpipe output and return
a well formatted protein_id column that includes both "leading protein ID"
and "Indistinguishable Proteins"
}
