% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe_legacy.R
\name{parse_fragpipe_proteins__legacy}
\alias{parse_fragpipe_proteins__legacy}
\title{THIS FUNCTION SHOULD ONLY BE USED FOR LEGACY FRAGPIPE DATASETS; PARSER FOR PROTEIN TABLES}
\usage{
parse_fragpipe_proteins__legacy(file_proteins)
}
\arguments{
\item{file_proteins}{full file path}
}
\description{
parse a proteins.tsv or combined_protein.tsv from fragpipe output and return
a well formatted protein_id column that includes both "leading protein ID"
and "Indistinguishable Proteins"
}
