% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{parse_fragpipe_proteins_from_combined_proteins}
\alias{parse_fragpipe_proteins_from_combined_proteins}
\title{Construct MS-DAP style protein group identifiers from a FragPipe 'combined_protein.tsv' file by appending all 'indistinguishable' proteins after the leading protein ID}
\usage{
parse_fragpipe_proteins_from_combined_proteins(filename)
}
\arguments{
\item{filename}{full path to a FragPipe 'combined_protein.tsv' file}
}
\description{
Construct MS-DAP style protein group identifiers from a FragPipe 'combined_protein.tsv' file by appending all 'indistinguishable' proteins after the leading protein ID
}
