% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe_ionquant.R
\name{parse_fragpipe_psm_ion_pair}
\alias{parse_fragpipe_psm_ion_pair}
\title{augment FragPipe ion.tsv with retention-time data from the respective psm.tsv file}
\usage{
parse_fragpipe_psm_ion_pair(file_ion, file_psm)
}
\arguments{
\item{file_ion}{full path to a ion.tsv file}

\item{file_psm}{full path to a psm.tsv file, this must match the \code{file_ion} argument (e.g. files in the same dir)}
}
\description{
note that this still leaves retention-time information for MBR hits, this data is not available in FragPipe output atm.
}
