% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{path_clean_slashes}
\alias{path_clean_slashes}
\title{regex replace back slashes and redundant forward/back-slashes with a single forward slash}
\usage{
path_clean_slashes(f)
}
\arguments{
\item{f}{full path to some file}
}
\description{
regex replace back slashes and redundant forward/back-slashes with a single forward slash
}
