% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{path_exists}
\alias{path_exists}
\title{check if a path exists}
\usage{
path_exists(path, file)
}
\arguments{
\item{path}{a directory on this computer. eg; "C:/temp" on windows or "/home/user1" on unix}

\item{file}{filename within this directory}
}
\description{
check if a path exists
}
