% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{pdf_combine_chunks}
\alias{pdf_combine_chunks}
\title{placeholder title
the function pdf_combine() from pdftools/qpdf package has issues (on windows) when combining hundreds of plots. error: "Too many open files"
so we here create a wrapper where we combine source files in batches
some references; https://stackoverflow.com/questions/40810704/error-with-r-function-download-file-too-many-open-files}
\usage{
pdf_combine_chunks(input, output)
}
\arguments{
\item{input}{todo}

\item{output}{todo}
}
\description{
placeholder title
the function pdf_combine() from pdftools/qpdf package has issues (on windows) when combining hundreds of plots. error: "Too many open files"
so we here create a wrapper where we combine source files in batches
some references; https://stackoverflow.com/questions/40810704/error-with-r-function-download-file-too-many-open-files
}
