% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_excel.R
\name{peptide_abundance_table}
\alias{peptide_abundance_table}
\title{placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))}
\usage{
peptide_abundance_table(
  peptides,
  proteins,
  samples,
  intensity_column_name = "intensity",
  norm_algorithm = ""
)
}
\arguments{
\item{peptides}{todo}

\item{proteins}{todo}

\item{samples}{todo}

\item{intensity_column_name}{todo}

\item{norm_algorithm}{todo}
}
\description{
placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))
}
