% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{peptides_collapse_by_sequence}
\alias{peptides_collapse_by_sequence}
\title{placeholder title}
\usage{
peptides_collapse_by_sequence(peptides, prop_peptide = "sequence_plain")
}
\arguments{
\item{peptides}{todo}

\item{prop_peptide}{todo}
}
\description{
placeholder title
}
