% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{permute_ab}
\alias{permute_ab}
\title{basic between-group permutation implementation}
\usage{
permute_ab(x, nmax = 100)
}
\arguments{
\item{x}{integer vector of {1,2} group IDs, at least two of each}

\item{nmax}{maximum number of permuted sets}
}
\value{
integer matrix where rows are permutations and columns are indices in input x
}
\description{
minimal code & dependency-free. Not computationally efficient, but this is negligible for our use-cases
}
\details{
permute_ab(as.integer(c(1,1,1, 2,2,2))) # most basic
permute_ab(as.integer(c(1,1,1,1, 2,2,2,2,2))) # uneven groups
}
