% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark.R
\name{plot_benchmark_volcano}
\alias{plot_benchmark_volcano}
\title{placeholder title
pre-define color_code and minlog10pval}
\usage{
plot_benchmark_volcano(tib, mtitle = "", color_discrete = TRUE)
}
\arguments{
\item{tib}{todo}

\item{mtitle}{todo}

\item{color_discrete}{todo}
}
\description{
placeholder title
pre-define color_code and minlog10pval
}
