% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_distributions.R
\name{plot_dia_cscore_histograms}
\alias{plot_dia_cscore_histograms}
\title{placeholder title}
\usage{
plot_dia_cscore_histograms(tib_input)
}
\arguments{
\item{tib_input}{todo}
}
\description{
placeholder title
}
