% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_foldchanges}
\alias{plot_foldchanges}
\title{placeholder title}
\usage{
plot_foldchanges(tib, mtitle = "")
}
\arguments{
\item{tib}{pre-define color_code column}

\item{mtitle}{todo}
}
\description{
placeholder title
}
