% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{plot_peptide_data}
\alias{plot_peptide_data}
\title{iterate all contrasts: gather peptide data used in upstream stats, create individual protein plots, finally merge plots into end result PDFs
temp filenames not random, so only run a single session of this tool at the same time on a given system}
\usage{
plot_peptide_data(
  peptides,
  proteins,
  samples,
  de_proteins,
  output_dir,
  plot_each_statistical_approach = TRUE,
  plot_all_proteins = FALSE,
  plot_comparison_msqrob_msempire = FALSE
)
}
\arguments{
\item{peptides}{todo}

\item{proteins}{proteins tibble needs columns: (protein_id, fasta_headers)}

\item{samples}{todo}

\item{de_proteins}{todo}

\item{output_dir}{todo}

\item{plot_each_statistical_approach}{todo}

\item{plot_all_proteins}{todo}

\item{plot_comparison_msqrob_msempire}{todo}
}
\description{
iterate all contrasts: gather peptide data used in upstream stats, create individual protein plots, finally merge plots into end result PDFs
temp filenames not random, so only run a single session of this tool at the same time on a given system
}
