% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{plot_peptide_data_by_contrast_to_pdf}
\alias{plot_peptide_data_by_contrast_to_pdf}
\title{plot peptide-level data}
\usage{
plot_peptide_data_by_contrast_to_pdf(
  dataset,
  filter_protein_id,
  output_dir,
  show_unused_datapoints = FALSE,
  norm_algorithm = NA,
  rollup_algorithm = "maxlfq"
)
}
\arguments{
\item{dataset}{todo}

\item{filter_protein_id}{todo}

\item{output_dir}{todo}

\item{show_unused_datapoints}{todo}

\item{norm_algorithm}{todo}

\item{rollup_algorithm}{todo}
}
\description{
plot peptide-level data
}
