% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{plot_peptide_data_per_protein}
\alias{plot_peptide_data_per_protein}
\title{placeholder title
proteins tibble needs columns: (protein_id, fasta_headers)}
\usage{
plot_peptide_data_per_protein(
  peptides,
  proteins,
  stats_de,
  samples,
  protein_ids
)
}
\arguments{
\item{peptides}{todo}

\item{proteins}{todo}

\item{stats_de}{todo}

\item{samples}{todo}

\item{protein_ids}{todo}
}
\description{
placeholder title
proteins tibble needs columns: (protein_id, fasta_headers)
}
