% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_pvalue_histogram}
\alias{plot_pvalue_histogram}
\title{placeholder title}
\usage{
plot_pvalue_histogram(tib, mtitle = "")
}
\arguments{
\item{tib}{todo}

\item{mtitle}{todo}
}
\description{
placeholder title
}
