% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark.R
\name{plot_roc}
\alias{plot_roc}
\title{placeholder title
prior to this; species %in% c("YEAST", "HUMAN")
add extra column predictor; true for spike-in, false for rest}
\usage{
plot_roc(tib, mtitle = "", universe = "all", plot_coords = FALSE)
}
\arguments{
\item{tib}{todo}

\item{mtitle}{todo}

\item{universe}{todo}

\item{plot_coords}{todo}
}
\description{
placeholder title
prior to this; species %in% c("YEAST", "HUMAN")
add extra column predictor; true for spike-in, false for rest
}
