% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_pca.R
\name{plot_sample_pca}
\alias{plot_sample_pca}
\title{placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))}
\usage{
plot_sample_pca(
  matrix_sample_intensities,
  samples,
  samples_colors,
  label_samples_by_shortname = TRUE,
  pch_as_exclude = TRUE
)
}
\arguments{
\item{matrix_sample_intensities}{todo}

\item{samples}{todo}

\item{samples_colors}{todo}

\item{label_samples_by_shortname}{todo}

\item{pch_as_exclude}{todo}
}
\description{
placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))
}
