% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_pca.R
\name{plot_sample_pca__sample_in_contrast}
\alias{plot_sample_pca__sample_in_contrast}
\title{helper function for use-case beyond making a report using standard analysis_quickstart() workflow; PCA on the subset of samples-of-interest from some contrast}
\usage{
plot_sample_pca__sample_in_contrast(
  dataset,
  contr,
  sample_label_property = "auto"
)
}
\arguments{
\item{dataset}{your dataset after application of dea. eg; results from analysis_quickstart()}

\item{contr}{the exact name of the contrast. Example; "contrast: ctrl1,ctrl2 vs phenotype1,phenotype2,phenotype3"  This should be one of the columns in dataset$samples table (which are generated by the setup_contrasts() function)}

\item{sample_label_property}{see plot_sample_pca() function for params}
}
\description{
helper function for use-case beyond making a report using standard analysis_quickstart() workflow; PCA on the subset of samples-of-interest from some contrast
}
