% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark.R
\name{plot_true_false_positive_counts}
\alias{plot_true_false_positive_counts}
\title{placeholder title}
\usage{
plot_true_false_positive_counts(tib, mtitle = "")
}
\arguments{
\item{tib}{todo}

\item{mtitle}{todo}
}
\description{
placeholder title
}
