% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance_explained.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot the variance explained of sample metadata properties in the protein-intensity matrix}
\usage{
plot_variance_explained(dataset, cols_metadata = NULL)
}
\arguments{
\item{dataset}{your dataset}

\item{cols_metadata}{columns from dataset@samples to be used. leave empty to automatically infer (default)}
}
\description{
basically a wrapper around the variancePartition R package (Hoffman GE, Schadt EE, 2016, PMID:27884101)
}
\details{
note; this is quite slow even on small datasets
}
