% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{placeholder title}
\usage{
plot_volcano(
  stats_de,
  log2foldchange_threshold = NA,
  qvalue_threshold = NA,
  mtitle = ""
)
}
\arguments{
\item{stats_de}{todo}

\item{log2foldchange_threshold}{todo}

\item{qvalue_threshold}{todo}

\item{mtitle}{todo}
}
\description{
placeholder title
}
