% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano_allcontrast}
\alias{plot_volcano_allcontrast}
\title{a simple wrapper around plot_volcano() that iterates all contrasts}
\usage{
plot_volcano_allcontrast(stats_de, mtitle, ...)
}
\arguments{
\item{stats_de}{input data.frame, see \code{plot_volcano()}}

\item{mtitle}{plot title, see \code{plot_volcano()}}

\item{...}{parameters passed to \code{plot_volcano()}, check that function's documentation and examples}
}
\description{
returns the plot_volcano() output within a list (1 element per contrast @ stats_de$contrast column)
}
\examples{
\dontrun{
  # refer to the examples for msdap::plot_volcano()
  # the same syntax can be applied here
}
}
