% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{print_available_filtering_results}
\alias{print_available_filtering_results}
\title{Return column names in peptide table that hold intensity values}
\usage{
print_available_filtering_results(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
Return column names in peptide table that hold intensity values
}
