% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{print_contrasts}
\alias{print_contrasts}
\title{Print an overview of all defined contrasts and their respective sample metadata tables to the console}
\usage{
print_contrasts(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
Print an overview of all defined contrasts and their respective sample metadata tables to the console
}
\seealso{
\code{\link[=add_contrast]{add_contrast()}}, \code{\link[=remove_contrasts]{remove_contrasts()}}
}
