% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{print_dataset_summary}
\alias{print_dataset_summary}
\title{Print a short summary of a dataset to console}
\usage{
print_dataset_summary(dataset)
}
\arguments{
\item{dataset}{dataset to print}
}
\description{
Print a short summary of a dataset to console
}
