% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{protein2gene_orthologs}
\alias{protein2gene_orthologs}
\title{Map the protein_id column, which are assumed to contain uniprot mouse/rat identifiers, in a table to HGNC human gene IDs}
\usage{
protein2gene_orthologs(x, hgnc, idmap, remove_nohgnc = FALSE)
}
\arguments{
\item{x}{a data.table with columns protein_id and symbol (e.g. the "leading symbol" extracted from the uniprot gene symbols @ dataset$proteins$gene_symbols_or_id)}

\item{hgnc}{HGNC lookup table from \code{hgnc_lookuptable()}}

\item{idmap}{MGI lookup table from \code{mgi_lookuptable()} or a RGD lookup table from \code{rgd_lookuptable()}}

\item{remove_nohgnc}{set to TRUE to remove all rows that could not be matched to a HGNC ID. default; FALSE}
}
\description{
This function is currently limited such that the first matching human gene per proteingroup is returned,
so ambiguous proteingroups with multiple gene symbols (e.g. "GRIA1;GRIA2") might not always yield the 'leading'
gene and additional/ambiguous genes are disregarded). Columns describing the HGNC (genenames.org) ID,
NCBI Entez ID and Ensembl gene ID are appended to the provided table.
}
\details{
Workflow for ID mapping, where subsequent steps are fallbacks;
\enumerate{
\item uniprot id -> mgi/rgd id -> hgnc id
\item uniprot symbol -> mgi/rgd id -> hgnc id
\item uniprot symbol -> hgnc official symbol -> hgnc id
\item uniprot symbol -> hgnc non-ambiguous synonyms -> hgnc id
}
}
\examples{
\dontrun{
  # prior to this, import a dataset/samples/fasta & setup contrasts & apply filter_datasets()
  # alternatively, use analysis_quickstart() to do everything
  dataset = dea(dataset, dea_algorithm = "deqms")
  dataset = differential_detect(dataset, min_peptides_observed = 2, min_samples_observed = 3,
                                min_fraction_observed = 0.5, return_wide_format = FALSE)
  print_dataset_summary(dataset)

  x = summarise_stats(dataset, return_dea = TRUE, return_diffdetect = TRUE,
                      diffdetect_zscore_threshold = 5, remove_ambiguous_proteingroups = TRUE)
  hgnc = hgnc_lookuptable(f = "C:/DATA/hgnc/hgnc__non_alt_loci_set___2023-03-21.tsv")
  mgi = mgi_lookuptable(f = "C:/DATA/mgi/2023-07-15/MRK_SwissProt_TrEMBL.rpt")
  y = protein2gene_orthologs(x, hgnc, mgi)
}
}
