% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{protein_eset_from_data}
\alias{protein_eset_from_data}
\title{Construct an ExpressionSet}
\usage{
protein_eset_from_data(x, eset)
}
\arguments{
\item{x}{protein abundance matrix}

\item{eset}{A Biobase ExpressionSet with protein_id and sample_id metadata}
}
\description{
Construct an ExpressionSet
}
