% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{protein_foldchange_from_ebayes}
\alias{protein_foldchange_from_ebayes}
\title{Compute protein log2 foldchanges from protein-level eBayes (i.e. simple, no shrinkage)}
\usage{
protein_foldchange_from_ebayes(
  eset,
  input_intensities_are_log2 = TRUE,
  random_variables = NULL
)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the protein intensities. required attributes; fData(): protein_id and pData(): condition}

\item{input_intensities_are_log2}{whether the provided intensities are already log2 transformed}

\item{random_variables}{a vector of column names in your sample metadata table that are added as additional(!) regression terms}
}
\value{
data.frame with columns protein_id and foldchange.log2
}
\description{
implementation analogous to \code{de_ebayes()}
}
