% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fasta.R
\name{proteingroup_to_idshort_lookuptable}
\alias{proteingroup_to_idshort_lookuptable}
\title{lookup table for proteingroup identifiers (in long or short ID form) to idshort and isdecoy}
\usage{
proteingroup_to_idshort_lookuptable(x)
}
\arguments{
\item{x}{array of proteingroup identifiers}
}
\description{
lookup table for proteingroup identifiers (in long or short ID form) to idshort and isdecoy
}
\examples{
\dontrun{
proteingroup_to_idshort_lookuptable(c(
  NA, "", # empty values are skipped
  "P36578", "P36578", "P36578", "P36578", # unique values are returned
  "P55011-3;P55011",
  "Q5TF21;rev_E9PJP2", "decoy_Q5TF21;E9PJP2",
  "sp|P63027|VAMP2_HUMAN;tr|K7ENK9|K7ENK9_HUMAN",
  "sp|P63027|VAMP2_HUMAN;rev_tr|K7ENK9|K7ENK9_HUMAN",
  "rev_sp|P63027|VAMP2_HUMAN;tr|K7ENK9|K7ENK9_HUMAN"
))
}
}
