% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{read_table_by_header_spec}
\alias{read_table_by_header_spec}
\title{Efficiently read a table based on some specification of expected headers}
\usage{
read_table_by_header_spec(
  file,
  attributes_required,
  attributes_optional = NULL,
  regex_headers = NULL,
  as_tibble_type = TRUE
)
}
\arguments{
\item{file}{full file path for a CSV/TSV data table}

\item{attributes_required}{see \code{map_headers()}}

\item{attributes_optional}{see \code{map_headers()}}

\item{regex_headers}{regex to be applied to all columns; these will be included}

\item{as_tibble_type}{return data.table::fread() results as a tibble (DEFAULT), instead of data.table format}
}
\description{
Uses map_headers() to parse and map lists of expected column names and their designated names
}
