% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{regex_classification}
\alias{regex_classification}
\title{Classify an array of strings by regular expression}
\usage{
regex_classification(x, regex, discard_label = "discard")
}
\arguments{
\item{x}{array of string where each regex is applied to}

\item{regex}{named vector of regular expressions. setting proper names is important, these are the classification labels. example: c(human="_HUMA", yeast="_YEAS", discard="_ECOL")}

\item{discard_label}{default label for classifications that should be discarded downstream. default: "discard"}
}
\description{
For example, provide fasta headers and an array of regular expressions to find out which are yeast/human while removing ambigous entries or those matching ecoli:
regex = c(human="_HUMA", yeast="_YEAS", discard="_ECOL")
}
