% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{regex_rawfile_strip_extension}
\alias{regex_rawfile_strip_extension}
\title{returns a regex string that can be used to remove extensions from mass-spec raw files}
\usage{
regex_rawfile_strip_extension()
}
\description{
returns a regex string that can be used to remove extensions from mass-spec raw files
}
