% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{remove_by_charlength}
\alias{remove_by_charlength}
\title{remove NA, non-character and characters shorter than N from array}
\usage{
remove_by_charlength(x, minchar)
}
\arguments{
\item{x}{character array}

\item{minchar}{integer, minimum character of elements in x}
}
\description{
remove NA, non-character and characters shorter than N from array
}
