% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{remove_contrasts}
\alias{remove_contrasts}
\title{Remove all contrast definitions and respective filtering and DEA results}
\usage{
remove_contrasts(dataset)
}
\arguments{
\item{dataset}{your dataset}
}
\description{
also removes contrast-related information from the dataset that was used in older MS-DAP version (prior to version 1.2)
}
\seealso{
\code{\link[=print_contrasts]{print_contrasts()}}, \code{\link[=add_contrast]{add_contrast()}}
}
