% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{remove_file_extension_from_path}
\alias{remove_file_extension_from_path}
\title{remove a single file extension from a path or filename}
\usage{
remove_file_extension_from_path(f)
}
\arguments{
\item{f}{f can either be a filename or a full path, and optionally carry a file extension. eg; remove_file_extension_from_path("file.txt"); remove_file_extension_from_path("C:/temp/file")}
}
\description{
purposely naive approach to keep things simple for our internal usecases; strip a . followed by 0-5 characters from end of the string
}
