% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{remove_file_if_exists}
\alias{remove_file_if_exists}
\title{remove file if exists}
\usage{
remove_file_if_exists(f)
}
\arguments{
\item{f}{full path to some file}
}
\description{
remove file if exists
}
