% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{replace_inf_with_NA}
\alias{replace_inf_with_NA}
\title{fast replacement in a data.frame; infinite values to NA
https://stackoverflow.com/a/57092321}
\usage{
replace_inf_with_NA(df)
}
\arguments{
\item{df}{data.frame to replace values in}
}
\description{
fast replacement in a data.frame; infinite values to NA
https://stackoverflow.com/a/57092321
}
