% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{replace_nonfinite}
\alias{replace_nonfinite}
\title{fast replacement in a vector}
\usage{
replace_nonfinite(x, val = NA)
}
\arguments{
\item{x}{vector to replace values in}

\item{val}{the value to replace with}
}
\description{
fast replacement in a vector
}
