% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_idmapping.R
\name{rgd_lookuptable}
\alias{rgd_lookuptable}
\title{Parse RGD ID mapping table table into a lookup table with symbols and uniprot IDs}
\usage{
rgd_lookuptable(f)
}
\arguments{
\item{f}{full path to the downloaded table (expected to be tsv format)}
}
\description{
download link: https://download.rgd.mcw.edu/data_release/GENES_RAT.txt
README: https://download.rgd.mcw.edu/data_release/GENES_README
}
