% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rmarkdown.R
\name{rmarkdown_xtable_custom}
\alias{rmarkdown_xtable_custom}
\title{placeholder title
https://stackoverflow.com/a/16580525}
\usage{
rmarkdown_xtable_custom(x, caption = NULL, align = NULL, scalebox = 0.9)
}
\arguments{
\item{x}{matrix/data.frame}

\item{caption}{todo}

\item{align}{todo}

\item{scalebox}{todo}
}
\description{
placeholder title
https://stackoverflow.com/a/16580525
}
