% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{robust_header_matching}
\alias{robust_header_matching}
\title{robust matching of column names to a data.frame, returns the matched subset of columns}
\usage{
robust_header_matching(x, column_spec, columns_required)
}
\arguments{
\item{x}{a data.frame}

\item{column_spec}{named list where names are the target column names and values are vectors of column names to search for in parameter x}

\item{columns_required}{a vector of columns that are required (i.e. throw error if any of these are missing). Should match the names in column_spec}
}
\description{
robust matching of column names to a data.frame, returns the matched subset of columns
}
