% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_pep2prot}
\alias{rollup_pep2prot}
\title{wrapper function for peptide-to-protein rollup}
\usage{
rollup_pep2prot(
  tib,
  intensity_is_log2 = TRUE,
  rollup_algorithm,
  return_as_matrix = FALSE
)
}
\arguments{
\item{tib}{peptide tibble in long-format. missing values should be set NA, not zero. required columns: peptide_id, protein_id, sample_id, intensity}

\item{intensity_is_log2}{boolean indicating if input peptide intensities are log2 scaled}

\item{rollup_algorithm}{algorithm for combining peptides to proteins. Options: 'maxlfq', 'sum', 'tukey_median'. The former applies the MaxLFQ algorithm, the second employs the 'classic' strategy of summing all peptides per protein (non-log values), the latter uses Tukey's median polish algorithm. Alternatively, one can specify "maxlfq_diann" to use the DIA-NN R package's implementation for MaxLFQ instead of the faster implementation from the iq R package (default when using 'maxlfq')}

\item{return_as_matrix}{instead of long-format tibble, return a protein*sample matrix (missing values set to NA)}
}
\description{
converts a table with peptide-level abundance values to protein-level data by combining data from respective peptides per protein (per sample).
}
