% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_pep2prot_maxlfq}
\alias{rollup_pep2prot_maxlfq}
\title{rollup by MaxLFQ algorithm}
\usage{
rollup_pep2prot_maxlfq(
  tib,
  intensity_is_log2,
  implementation = "iq",
  return_as_matrix = FALSE
)
}
\arguments{
\item{tib}{peptide tibble in long-format. missing values should be set NA, not zero. required columns: peptide_id, protein_id, sample_id, intensity}

\item{intensity_is_log2}{boolean indicating if input peptide intensities are log2 scaled}

\item{implementation}{implementation to use. available options: "iq", "diann"}

\item{return_as_matrix}{instead of long-format tibble, return a protein*sample matrix (missing values set to NA)}
}
\description{
Available implementations:
iq R package: https://github.com/tvpham/iq/
DIA-NN R package: https://github.com/vdemichev/DiaNN/
}
