% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_pep2prot_summation}
\alias{rollup_pep2prot_summation}
\title{rollup by plain peptide intensity summation}
\usage{
rollup_pep2prot_summation(
  tib,
  intensity_is_log2 = TRUE,
  return_as_matrix = FALSE
)
}
\arguments{
\item{tib}{peptide tibble in long-format. missing values should be set NA, not zero. required columns: peptide_id, protein_id, sample_id, intensity}

\item{intensity_is_log2}{boolean indicating log2 scaling}

\item{return_as_matrix}{instead of long-format tibble, return a protein*sample matrix (missing values set to NA)}
}
\description{
rollup by plain peptide intensity summation
}
