% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_pep2prot_tmp}
\alias{rollup_pep2prot_tmp}
\title{rollup a peptide tibble by applying Tukey's median polish algorithm to all peptides per protein}
\usage{
rollup_pep2prot_tmp(tib, intensity_is_log2 = TRUE, return_as_matrix = FALSE)
}
\arguments{
\item{tib}{peptide tibble in long-format. Missing values should be set NA, not zero. required columns: peptide_id, protein_id, sample_id, intensity}

\item{intensity_is_log2}{boolean indicating if input peptide intensities are log2 scaled}

\item{return_as_matrix}{instead of long-format tibble, return a protein*sample matrix (missing values set to NA)}
}
\description{
rollup a peptide tibble by applying Tukey's median polish algorithm to all peptides per protein
}
