% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_metadata.R
\name{sample_color_coding}
\alias{sample_color_coding}
\title{generate color-coding for all sample metadata}
\usage{
sample_color_coding(samples)
}
\arguments{
\item{samples}{sample metadata table, typically; dataset$samples}
}
\description{
generate color-coding for all sample metadata
}
