% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_metadata.R
\name{sample_color_coding_as_long_tibble}
\alias{sample_color_coding_as_long_tibble}
\title{convert color matrix to long format}
\usage{
sample_color_coding_as_long_tibble(samples, samples_colors)
}
\arguments{
\item{samples}{todo}

\item{samples_colors}{todo}
}
\description{
convert color matrix to long format
}
