% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata_custom}
\alias{sample_metadata_custom}
\title{For advanced users: import sample metadata directly from the sample names using regular expressions}
\usage{
sample_metadata_custom(
  dataset,
  sample_property_regex = list(),
  sample_exclude_regex = "",
  group_order = NA,
  group_regex_array = NA
)
}
\arguments{
\item{dataset}{your dataset}

\item{sample_property_regex}{a named list of regular expressions to extract sample properties. name=sample property, value = array of length two with the regex and the replace string (first 2 arguments of sub())}

\item{sample_exclude_regex}{optionally, regex to select samples that should be flagged as 'exclude'}

\item{group_order}{optionally, you can provide a preferred ordering of the sample groups as a character array}

\item{group_regex_array}{optionally, a regex that specifically extracts sample groups and leaves non-matches in an 'unassigned' group}
}
\description{
For advanced users: import sample metadata directly from the sample names using regular expressions
}
\examples{
# example of a regex applied to sample names to extract the sample shortname and group
\dontrun{sample_property_regex = list(
  shortname = c(".*experimentname_([A-Z]+.\\\\d).*", "\\\\1"),
  group = c(".*experimentname_([A-Z]+)_.*", "\\\\1")
  )
}
}
