% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata_sort_and_filter}
\alias{sample_metadata_sort_and_filter}
\title{sample metadata formatting and validation}
\usage{
sample_metadata_sort_and_filter(
  df,
  sample_exclude_regex = "",
  group_order = NA
)
}
\arguments{
\item{df}{data.frame with sample metadata}

\item{sample_exclude_regex}{optionally, regex to select samples that should be flagged as 'exclude'}

\item{group_order}{optionally, you can provide a preferred ordering of the sample groups as a character array}
}
\description{
sample metadata formatting and validation
}
