% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{scale_between_quantiles}
\alias{scale_between_quantiles}
\title{scale an array of numerics between 0~1, thresholding values at lower/upper quantiles
non-finite values are ignored (not replaced).
example: x=rnorm(1000); hist(x); hist(scale_between_quantiles(x, min_quantile=0.01, max_quantile = 0.99))}
\usage{
scale_between_quantiles(x, min_quantile = 0.01, max_quantile = 0.99)
}
\arguments{
\item{x}{values to transform}

\item{min_quantile}{lower quantile, used as parameter in quantile()}

\item{max_quantile}{upper quantile, used as parameter in quantile()}
}
\description{
edge-case; if either lower or upper quantile could not be determined, all finite values of x will be set to 1
edge-case; if all non-finite input values are the same, the upper and lower quantiles are the same. All finite values of x will be set to 1
}
