% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{score_classification_counts}
\alias{score_classification_counts}
\title{score which classification works best when e.g. guesstimating which columns to split}
\usage{
score_classification_counts(x)
}
\arguments{
\item{x}{character array or matrix}
}
\description{
score which classification works best when e.g. guesstimating which columns to split
}
