% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{shift_empty}
\alias{shift_empty}
\title{try to align matrix columns
eg; inconsistent use of separation characters leads to misalignment. neuron_glia_1, something_1 -->> should have been neuron-glia_1}
\usage{
shift_empty(x)
}
\arguments{
\item{x}{character matrix}
}
\description{
try to align matrix columns
eg; inconsistent use of separation characters leads to misalignment. neuron_glia_1, something_1 -->> should have been neuron-glia_1
}
