% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{shiny_show_screen}
\alias{shiny_show_screen}
\title{shiny function to show/hide elements}
\usage{
shiny_show_screen(idshow, idall)
}
\arguments{
\item{idshow}{list of all ids (to hide)}

\item{idall}{single id to show after hiding all}
}
\description{
shiny function to show/hide elements
}
