% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{simple_rollup_peptide_matrix}
\alias{simple_rollup_peptide_matrix}
\title{placeholder title}
\usage{
simple_rollup_peptide_matrix(x_as_log2, grp_var)
}
\arguments{
\item{x_as_log2}{todo}

\item{grp_var}{todo}
}
\description{
placeholder title
}
