% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{split_alt_sep_char}
\alias{split_alt_sep_char}
\title{detect case of forgotten separation character, alternative chars}
\usage{
split_alt_sep_char(x, sep_chars)
}
\arguments{
\item{x}{array of characters}

\item{sep_chars}{array of separation characters}
}
\description{
detect case of forgotten separation character, alternative chars
}
