% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{split_char_num}
\alias{split_char_num}
\title{detect case of forgotten separation character}
\usage{
split_char_num(x)
}
\arguments{
\item{x}{array of characters}
}
\description{
examples; wt1, wt2  ,  1plus, 1min, 2plus, 2min
}
