% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_infer_metadata_from_file.R
\name{split_recursively}
\alias{split_recursively}
\title{recursively split within all columns of a matrix}
\usage{
split_recursively(x, alt_sep_chars)
}
\arguments{
\item{x}{character matrix}

\item{alt_sep_chars}{array of separation characters}
}
\description{
recursively split within all columns of a matrix
}
