% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{strip_common_substring}
\alias{strip_common_substring}
\title{remove starting or trailing substrings that ar the same for all strings in a vector}
\usage{
strip_common_substring(s)
}
\arguments{
\item{s}{array of strings}
}
\description{
eg; \code{.txt} in all filesnames, or \verb{<experiment data><samplename>}
}
\examples{
\dontrun{ strip_common_substring(s = unique(peptides$sample_id)) }
\dontrun{ strip_common_substring(s = c("sample_a.txt","sample_b1.txt","sample_xyz.txt")) }
\dontrun{ strip_common_substring(s = c("","sample_b1.txt","sample_xyz.txt")) }
\dontrun{ strip_common_substring(s = c("sample.txt","sample_b1.txt","sample_xyz.txt")) }
\dontrun{
  assertthat::are_equal(
    strip_common_substring(c("sample_a.txt","sample_b1.txt","sample_xyz.txt")),
    c("a","b1","xyz")
  )
}

}
