% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rmarkdown.R
\name{subchunkify}
\alias{subchunkify}
\title{placeholder title
https://github.com/yihui/knitr/issues/1494
http://michaeljw.com/blog/post/subchunkify/}
\usage{
subchunkify(g, unique_chunk_id, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{todo}

\item{unique_chunk_id}{todo}

\item{fig_height}{todo}

\item{fig_width}{todo}
}
\description{
placeholder title
https://github.com/yihui/knitr/issues/1494
http://michaeljw.com/blog/post/subchunkify/
}
