% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{subset_eset}
\alias{subset_eset}
\title{placeholder title}
\usage{
subset_eset(eset, valid_peptide_ids, valid_sample_ids)
}
\arguments{
\item{eset}{todo}

\item{valid_peptide_ids}{todo}

\item{valid_sample_ids}{todo}
}
\description{
placeholder title
}
