% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{subset_protein_eset_for_dea}
\alias{subset_protein_eset_for_dea}
\title{helper function to double-check we don't submit an ExpressionSet with too few values to DEA}
\usage{
subset_protein_eset_for_dea(
  eset,
  model_matrix_result_prop,
  is_peptide,
  method_name
)
}
\arguments{
\item{eset}{a Biobase ExpressionSet}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}

\item{is_peptide}{boolean indicating peptide-level (FALSE = protein-level)}

\item{method_name}{label for printing error messages}
}
\description{
helper function to double-check we don't submit an ExpressionSet with too few values to DEA
}
