% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{subset_relevant_code_snippet_for_report}
\alias{subset_relevant_code_snippet_for_report}
\title{subset lines of R code (provided as strings) from the last "import_dataset" function on}
\usage{
subset_relevant_code_snippet_for_report(rcode_lines)
}
\arguments{
\item{rcode_lines}{R code as an array of characters (strings)}
}
\description{
in case users run the pipeline a few times consecutively (eg; testing various parameters) without resetting R history, we only want to keep the relevant portion of the code.
simple/transparent way is to show everything from latest dataset import function or source() on
}
