% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{threshold_numerics}
\alias{threshold_numerics}
\title{threshold finite numeric values}
\usage{
threshold_numerics(x, threshold)
}
\arguments{
\item{x}{numeric array or matrix}

\item{threshold}{limit finite values below this val}
}
\description{
threshold finite numeric values
}
