% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{tibble_as_eset}
\alias{tibble_as_eset}
\title{peptide-level ExpressionSet from long-format peptide data}
\usage{
tibble_as_eset(peptides, proteins, samples)
}
\arguments{
\item{peptides}{peptide tibble}

\item{proteins}{protein tibble}

\item{samples}{sample tibble}
}
\description{
peptide-level ExpressionSet from long-format peptide data
}
