% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{tibble_peptides_reorder}
\alias{tibble_peptides_reorder}
\title{rearrange column in a long-format peptides tibble}
\usage{
tibble_peptides_reorder(tib)
}
\arguments{
\item{tib}{peptide tibble in long format}
}
\description{
rearrange column in a long-format peptides tibble
}
