% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_excel.R
\name{tibble_prop_to_wide_and_align}
\alias{tibble_prop_to_wide_and_align}
\title{placeholder title}
\usage{
tibble_prop_to_wide_and_align(peptides, col, ref_metadata, ref_colnames)
}
\arguments{
\item{peptides}{todo}

\item{col}{todo}

\item{ref_metadata}{todo}

\item{ref_colnames}{todo}
}
\description{
placeholder title
}
