% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{toggle_inputs}
\alias{toggle_inputs}
\title{toggle shiny input elements}
\usage{
toggle_inputs(
  input_list,
  enable_components = T,
  exclude_names = c("log", "btnScreenPrevious", "btnScreenNext")
)
}
\arguments{
\item{input_list}{List of inputs, eg; input_list = reactiveValuesToList(input)}

\item{enable_components}{boolean flag; enable or disable?}

\item{exclude_names}{element names to exclude}
}
\description{
https://stackoverflow.com/a/48852284
}
