% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{unpack_list}
\alias{unpack_list}
\title{placeholder title
https://stackoverflow.com/a/40622857}
\usage{
unpack_list(lst, envir = parent.frame())
}
\arguments{
\item{lst}{todo}

\item{envir}{todo}
}
\description{
placeholder title
https://stackoverflow.com/a/40622857
}
