% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longformat_generic.R
\name{update_protein_mapping}
\alias{update_protein_mapping}
\title{For advanced users: update a peptide tibble with new peptide-to-protein mappings}
\usage{
update_protein_mapping(tib, pep2prot)
}
\arguments{
\item{tib}{peptides tibble, e.g.; dataset$peptides}

\item{pep2prot}{a peptide-to-protein mapping table, for instance one obtained from maxquant using import_maxquant_proteingroups()}
}
\description{
For advanced users: update a peptide tibble with new peptide-to-protein mappings
}
