% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longformat_generic.R
\name{update_protein_mapping_from_maxquant}
\alias{update_protein_mapping_from_maxquant}
\title{Update the peptide-to-protein mappings using a MaxQuant search result}
\usage{
update_protein_mapping_from_maxquant(
  dataset,
  path_maxquant,
  remove_shared = TRUE
)
}
\arguments{
\item{dataset}{your dataset}

\item{path_maxquant}{full path to MaxQuant output ('txt') folder}

\item{remove_shared}{remove all shared peptides that could not be assigned to a single proteingroup as a unique- or razor-peptide ?}
}
\description{
Simple wrapper function for user convenience; read MaxQuant proteinsgroups using import_maxquant_proteingroups(), then update the peptide-to-protein mappings in the peptide data table update_protein_mapping()
}
