% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{user_provided_metadata}
\alias{user_provided_metadata}
\title{Get column names that represent sample metadata provided by the user}
\usage{
user_provided_metadata(samples)
}
\arguments{
\item{samples}{e.g. dataset$samples}
}
\description{
excludes all MS-DAP standard columns, like sample_id and shortname but also
computed metadata like peptide or protein counts
}
