% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{validate_eset_fitdata}
\alias{validate_eset_fitdata}
\title{Input validation for eBayes/DEqMS/MS-EmpiRe functions}
\usage{
validate_eset_fitdata(eset, model_matrix, model_matrix_result_prop)
}
\arguments{
\item{eset}{protein/peptide log2 intensity matrix stored as a Biobase ExpressionSet. Must describe protein_id and sample_id in metadata}

\item{model_matrix}{a \code{stats::model.matrix()} result that is supplied to \code{limma::lmFit()}}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}
}
\description{
Input validation for eBayes/DEqMS/MS-EmpiRe functions
}
