% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_excel.R
\name{write_peptide_abundance_matrix_to_file}
\alias{write_peptide_abundance_matrix_to_file}
\title{Create an Excel document with all peptide abundances}
\usage{
write_peptide_abundance_matrix_to_file(
  dataset,
  filename,
  norm_algorithm = "",
  norm_modebetween_protein_eset = F
)
}
\arguments{
\item{dataset}{your dataset}

\item{filename}{full file path to output file}

\item{norm_algorithm}{normalization algorithm(s)}

\item{norm_modebetween_protein_eset}{normalise protein-level data matrix used for DEA by 'modebetween' to correct for 'imbalance' between conditions introduced in peptide-to-protein rollup. Note that some algo_de, such as msqrob or msempire, directly operate on the peptide-level data thus are unaffected by this setting. Only useful for statistical models that first roll-up to protein-level, such as eBayes. default:FALSE}
}
\description{
Multiple sheets indicate results from all filters applied. For large datasets this results in huge files, so only recommended for small datasets.
}
