% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_excel.R
\name{write_statistical_results_to_file}
\alias{write_statistical_results_to_file}
\title{write_statistical_results_to_file}
\usage{
write_statistical_results_to_file(dataset, output_dir)
}
\arguments{
\item{dataset}{your dataset}

\item{output_dir}{path to existing directory}
}
\description{
write_statistical_results_to_file
}
