% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{write_template_for_sample_metadata}
\alias{write_template_for_sample_metadata}
\title{Write an Excel sheet to file that contains all samples in the dataset}
\usage{
write_template_for_sample_metadata(dataset, filename, overwrite = FALSE)
}
\arguments{
\item{dataset}{your dataset}

\item{filename}{full file path of the desired output file (eg; C:/temp/template_experiment1.xlsx)}

\item{overwrite}{logical indicating whether to overwrite existing file (if any)}
}
\description{
Write an Excel sheet to file that contains all samples in the dataset
}
