% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstart.R
\name{analysis_quickstart}
\alias{analysis_quickstart}
\title{Quickstart for analyses in this pipeline}
\usage{
analysis_quickstart(
  dataset,
  filter_min_detect = 0,
  filter_fraction_detect = 0,
  filter_min_quant = 0,
  filter_fraction_quant = 0,
  filter_min_peptide_per_prot = 1,
  filter_topn_peptides = 0,
  filter_by_contrast = FALSE,
  norm_algorithm = c("vwmb", "modebetween_protein"),
  dea_protein_rollup = "maxlfq",
  dea_algorithm = c("deqms", "msqrob"),
  dea_qvalue_threshold = 0.01,
  dea_log2foldchange_threshold = 0,
  diffdetect_min_samples_observed = 3,
  plot_pca_label_by_shortname = NA,
  output_abundance_tables = FALSE,
  output_qc_report = TRUE,
  output_dir,
  output_within_timestamped_subdirectory = TRUE,
  dump_all_data = FALSE
)
}
\arguments{
\item{dataset}{a valid dataset object generated upstream by, for instance, import_dataset_skyline}

\item{filter_min_detect}{in order for a peptide to 'pass' in a sample group, in how many replicates must it be detected?}

\item{filter_fraction_detect}{in order for a peptide to 'pass' in a sample group, what fraction of replicates must it be detected?}

\item{filter_min_quant}{in order for a peptide to 'pass' in a sample group, in how many replicates must it be quantified?}

\item{filter_fraction_quant}{in order for a peptide to 'pass' in a sample group, what fraction of replicates must it be quantified?}

\item{filter_min_peptide_per_prot}{in order for a peptide to 'pass' in a sample group, how many peptides should be available after detect filters?}

\item{filter_topn_peptides}{maximum number of peptides to maintain for each protein (from the subset that passes above filters, peptides are ranked by the number of samples where detected and their variation between replicates). 1 is default, 2 can be a good choice situationally. If set to 1, make sure to inspect individual peptide data/plots for proteins with 1 peptide.}

\item{filter_by_contrast}{should the above filters be applied to all sample groups, or only those tested within each contrast? Enabling this optimizes available data in each contrast, but increases the complexity somewhat as different subsets of peptides are used in each contrast and normalization is applied separately.}

\item{norm_algorithm}{normalization algorithms. Can be empty string to skip normalization, or any algorithm listed by msdap::normalization_algorithms(). Provide an array of options to run each algorithm consecutively, for instance; c("vsn", "modebetween_protein") to first apply vsn normalization and then correct between-group ratios such that the protein-level log2-foldchange mode is zero}

\item{dea_protein_rollup}{strategy for combining peptides to proteins as used in DEA algorithms that first combine peptides to proteins and then apply statistics, like ebayes and deqms. options: maxlfq, sum. The former applies the MaxLFQ algorithm, the latter employs the 'classic' strategy of summing all peptides per protein}

\item{dea_algorithm}{algorithms for differential expression analysis. options: ebayes, deqms, msqrobsum, msempire, msqrob (to run multiple, provide an array)}

\item{dea_qvalue_threshold}{threshold for significance of adjusted p-values in figures and output tables}

\item{dea_log2foldchange_threshold}{threshold for significance of log2 foldchanges. Set to zero to disregard, a positive value to apply a cutoff to absolute foldchanges or use bootstrap analyses to infer a suitable foldchange threshold by providing either NA or a negative value. default: 0}

\item{diffdetect_min_samples_observed}{for differential detection only; minimum number of samples where a protein should be observed at least once by any of its peptides (in either group) when comparing a contrast of group A vs B}

\item{plot_pca_label_by_shortname}{whether to use sample names or a numeric ID as labels in the PCA plot. options: NA (let the code decide, default), TRUE (always use sample 'shortname'), FALSE (always use numeric ID)}

\item{output_abundance_tables}{whether to create an Excel document with all peptide abundances, with multiple sheets indicating results from all filters applied. For large datasets this results in huge files, so only recommended for small datasets. options: FALSE, TRUE}

\item{output_qc_report}{whether to create the Quality Control report. options: FALSE, TRUE . Highly recommended to set to TRUE. Set to FALSE to skip the report PDF (eg; to just do statistics and skip the time-consuming report creation)}

\item{output_dir}{output directory where all output files are stored, must be an existing directory}

\item{output_within_timestamped_subdirectory}{optionally, automatically create a subdirectory (without output_dir) that has the current date&time as name and store results there}

\item{dump_all_data}{if you're interested in performing custom bioinformatic analyses and want to use any of the data generated by this tool, you can dump all intermediate files to disk. Has performance impact so don't enable by default. options: FALSE, TRUE}
}
\description{
all-in-one function that covers the vast majority of use-cases. Typically, this function is all you need (besides importing your data upstream).
}
\section{Filtering}{


Peptide filter criteria applied to replicate samples within a sample group.
params; filter_min_detect, filter_fraction_detect, filter_min_quant, filter_fraction_quant.
You only have to provide active filters (but specify at least 1), filters/settings you do not specify don't do anything by default.

Settings:
for DDA: at least 1~2 detect (MS/MS ID) and quantified in at least ~75% of replicates.
for DIA: detect (confidence score < threshold) in at least ~75% of replicates (because for DIA, you typically have an abundance value in each sample regardless of the identifier confidence score).
If there are only 3 replicates, we recommend filtering such that there are at least 3 datapoints to work with in differential expression analysis.

Taken together, recommended settings for a DDA dataset with 3~8 replicates in each sample group look like this;

\code{
filter_min_detect = 1, filter_fraction_detect = 0.25, filter_min_quant = 3, filter_fraction_quant = 0.75
}

Analogous for DIA;

\code{
filter_min_detect = 3, filter_fraction_detect = 0.75
}
}

\section{Filter within contrast vs using all groups}{


Two distinct approaches to selecting peptides can be used for differential expression analysis: 1) 'within contrast' and 2) 'apply filter to all sample groups'.

1) Determine within each contrast (eg; group A vs group B) what peptides can be used by applying above peptide filter criteria and then apply normalization to this data subset.
Advantaguous in datasets with many groups; this maximizes the number of peptides used in each contrast (eg; let peptide p be observed in groups A and B, not in C. we'd want to use it in A vs B, not in A vs C).
As a disadvantage, this complicates interpretation since the exact data used is different in each contrast (slightly different peptides and normalization in each contrast).

2) Apply above filter criteria to each sample group (eg; a peptide must past these filter rules in every sample group) and then apply normalization

This data matrix is then used for all downstream statistics

Advantage; simple and robust

Disadvantage; potentially miss out on (group-specific) peptides/data-points that may fail filter criteria in just 1 group, particularly in large datasets with 4+ groups

Set \code{filter_within_contrast = FALSE} for this option

Note; if there are just 2 sample groups (eg; WT vs KO), this point is moot as both approaches are the same
}

\section{Normalization}{

options: "" (empty string disables normalization), "vsn", "loess", "rlr", "msempire", "vwmb", "modebetween".

You can combine normalizations by providing an array of options to apply subsequential normalizations.

For instance, \code{norm_algorithm = c("vsn", "modebetween")} applies the vsn algorithm (quite strong normalization reducing variation) and then balances between-group foldchanges with the modebetween normalization from the vwmb algorithm.

By default uses the built-in normalization algorithm Variation Within, Mode Between (vwmb).
Benchmarks have shown that vsn&modebetween (code example above) is a great alternative to consider.
}

\section{Differential Expression Analysis}{


Statistical models for differential expression analysis

MSqRob is recommended for most cases; a peptide-level model that is highly sensitive and quite robust. Reference: https://github.com/statOmics/MSqRob

MS-EmpiRe a peptide-level model that works especially well for DDA data. Reference: https://github.com/zimmerlab/MS-EmpiRe

eBayes is robust but conservative, using the limma package to apply moderated t-tests on protein-level abundances. Reference: https://doi.org/doi:10.18129/B9.bioc.limma

options: ebayes, msempire, msqrob (to run multiple, provide an array)

Users are encouraged to simply run all statistical models and consider the set of candidate proteins that pop up in multiple approaches. Output tables already contain columns to convenience this approach, if 3 DEA algorithms have been selected.
e.g.; \code{dea_algorithm = c("ebayes", "msempire", "msqrob")}

All statistical results are shown in the QC report and output tables and can be easily compared.
}

