% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{check_parameter_is_numeric}
\alias{check_parameter_is_numeric}
\title{placeholder title}
\usage{
check_parameter_is_numeric(..., minval = NA, maxval = NA)
}
\arguments{
\item{...}{todo}

\item{minval}{todo}

\item{maxval}{todo}
}
\description{
placeholder title
}
