% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_deqms}
\alias{de_deqms}
\title{Apply DEqMS to a protein-level ExpressionSet}
\usage{
de_deqms(eset, model_matrix, model_matrix_result_prop, doplot = FALSE)
}
\arguments{
\item{eset}{protein-level log2 intensity matrix stored as a Biobase ExpressionSet. Note that it must contain protein metadata column 'npep' that holds integer peptide counts}

\item{model_matrix}{a \code{stats::model.matrix()} result that is supplied to \code{limma::lmFit()}}

\item{model_matrix_result_prop}{the column name in \code{model_matrix} that should be returned as the resulting coefficient estimated by \code{eBayes()}. In MS-DAP workflows this is typically "condition"}

\item{doplot}{create a QC plot?}
}
\description{
ref; PMID:32205417
ref; https://github.com/yafeng/DEqMS
}
\details{
implementation follows example code from the vignette; https://bioconductor.org/packages/release/bioc/vignettes/DEqMS/inst/doc/DEqMS-package-vignette.html
}
