% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_deqms}
\alias{de_deqms}
\title{apply DEqMS to a protein-level ExpressionSet}
\usage{
de_deqms(
  eset_proteins,
  peptides,
  input_intensities_are_log2 = TRUE,
  random_variables = NULL,
  doplot = FALSE
)
}
\arguments{
\item{eset_proteins}{protein-level dataset stored as a Biobase ExpressionSet}

\item{peptides}{peptide-level data tibble, must contain a protein_id and peptide_id column, used to count the number of unique peptides per protein}

\item{input_intensities_are_log2}{boolean indicating whether the ExpressionSet's intensity values are already log2 scaled (default: TRUE)}

\item{random_variables}{a vector of column names in your sample metadata table that are added as additional(!) regression terms in each statistical contrast tested downstream}

\item{doplot}{create a QC plot?}
}
\description{
ref; PMID:32205417
ref; https://github.com/yafeng/DEqMS
}
\details{
implementation follows example code from the vignette; https://bioconductor.org/packages/release/bioc/vignettes/DEqMS/inst/doc/DEqMS-package-vignette.html
}
