% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_ebayes}
\alias{de_ebayes}
\title{Wrapper function for limma::ebayes()}
\usage{
de_ebayes(x, mask_sample_groups)
}
\arguments{
\item{x}{log2 transformed protein intensity matrix}

\item{mask_sample_groups}{numeric or string array indicating to which group each column/sample belongs. NO NA VALUES !}
}
\description{
rownames of the data matrix (log2 intensities) must be the protein_id for downstream compatibility
}
