% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msempire}
\alias{de_msempire}
\title{Wrapper function for msEmpiRe::de.ana()}
\usage{
de_msempire(eset, input_intensities_are_log2 = F)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the peptide intensities. required attributes; fData(): protein_id and pData(): condition}

\item{input_intensities_are_log2}{whether the provided intensities are already log2 transformed}
}
\description{
input data must NOT be log transformed
}
\details{
TODO: proper standarderror computation
}
