% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msempire}
\alias{de_msempire}
\title{MS-EmpiRe implementation, a wrapper function for msEmpiRe::de.ana()}
\usage{
de_msempire(eset, input_intensities_are_log2 = F)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the peptide intensities. required attributes; fData(): protein_id and pData(): condition}

\item{input_intensities_are_log2}{whether the provided intensities are already log2 transformed}
}
\description{
ref; PMID:31235637
ref; https://github.com/zimmerlab/MS-EmpiRe
}
\details{
input data should NOT be log2 transformed (if it is, make sure to set parameter \code{input_intensities_are_log2} so we can revert this prior to passing data to MS-EmpiRe)

note; MS-EmpiRe crashes when all proteins have the same number of peptides (eg; filtering topn=1 or topn=3 and minpep=3)

TODO: proper standarderror computation
}
