% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_abundance.R
\name{de_msqrobsum_msqrob}
\alias{de_msqrobsum_msqrob}
\title{Wrapper function for msqrob, using the implementation from the msqrobsum package}
\usage{
de_msqrobsum_msqrob(
  eset,
  use_peptide_model = T,
  input_intensities_are_log2 = T,
  protein_rollup_robust = T
)
}
\arguments{
\item{eset}{a Biobase ExpressionSet that contains the peptide intensities. required attributes; fData(): protein_id and pData(): condition}

\item{use_peptide_model}{if true, apply msqrob. if false, apply msqrobsum}

\item{input_intensities_are_log2}{whether the provided intensities are already log2 transformed}

\item{protein_rollup_robust}{for msqrobsum analyses, whether to use 'robust' protein rollup (TRUE) or traditional peptide intensity summation (FALSE)}
}
\description{
Wrapper function for msqrob, using the implementation from the msqrobsum package
}
