% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_protein_background_foldchange_limits}
\alias{dea_protein_background_foldchange_limits}
\title{estimate a threshold for 'significant' foldchanges from N permutations of sample-to-condition assignments}
\usage{
dea_protein_background_foldchange_limits(
  eset,
  probs = 0.95,
  max_permutations = 100
)
}
\arguments{
\item{eset}{an ExpressionSet (works with both proteins and peptides)}

\item{probs}{upper limit for the quantile cutoff, automatically translated to mirror the lower limit; c(1-probs, probs)}

\item{max_permutations}{maximum number of unique configurations used for the permuted datasets}
}
\description{
note; this function hardcodes set.seed()
}
\details{
Permutations of sample labels within a group are disregarded as these have no effect on the between-group foldchange, only unique combinations of swapping samples between conditions A and B are considered

This is somewhat similar to the method described by Hafemeister and Satija at https://doi.org/10.1186/s13059-019-1874-1
M&M quote from this reference: "A random background distribution of mean differences was generated by randomly choosing 1000 genes and permuting the group labels. Significance thresholds for the difference of means were derived from the background distribution by taking the 0.5th and 99.5th percentile."
}
