% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_detect.R
\name{differential_detect}
\alias{differential_detect}
\title{Compare the number of peptide detection counts between groups}
\usage{
differential_detect(dataset, min_samples_observed = 3)
}
\arguments{
\item{dataset}{a valid dataset}

\item{min_samples_observed}{minimum number of samples where a protein should be observed at least once by any of its peptides (in either group) when comparing a contrast of group A vs B}
}
\description{
Compute a z-score for each protein based on the total number of detected peptides per sample group.
}
\details{
To easily prioritize proteins-of-interest, some which may not have sufficient abundance values for differential expression analysis but that do have many more detects in one sample group than the other, we provide a simple score based on identification count data.

This is a simplified approach that is intended to rank proteins for further qualitative analysis, be careful of over-interpretation and keep differences in sample group size (#replicates) and the absolute amount of peptides identified in a sample in mind !
}
