% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_differential_detect.R
\name{differential_detect}
\alias{differential_detect}
\title{Experimental feature: compare the number of peptide detection counts between groups}
\usage{
differential_detect(dataset)
}
\arguments{
\item{dataset}{a valid dataset}
}
\description{
To easily prioritize proteins-of-interest which do not have sufficient abundance values for statistical analysis, but that do have many more detects in one sample group than the other, we provide a simple score based on the count data from the preceding columns.
}
\details{
This is an intentionally simplified approach that only serves to rank proteins for further qualitative analysis.

First, adjust the peptide detects per sample by the the total number in the sample to account for systematic differences.

The zscore metric in diff_detect_zscore_contrast: <contrast>'; A='count_peptides_detected_within_group_A';  B='count_peptides_detected_within_group_A';  result = scale(log2(B+min(Bi)) - log2(A+min(Ai))).    (where scale = subtract the mean, then divide by standard deviation)

For all differential detection counts, be careful of over-interpretation and keep differences in sample group size (#replicates) and the total amount of peptides identified in a sample in mind !
}
