% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fasta.R
\name{fasta_header_to_gene}
\alias{fasta_header_to_gene}
\title{extract gene symbol from fasta header string}
\usage{
fasta_header_to_gene(x, fasta_id_type = "uniprot")
}
\arguments{
\item{x}{array of fasta headers}

\item{fasta_id_type}{fasta type. unused argument atm}
}
\description{
extract gene symbol from fasta header string
}
