% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fasta.R
\name{fasta_parse_file}
\alias{fasta_parse_file}
\title{reasonably fast method for reading header rows from all fasta files (without over-complicating code/implementation to improve by a mere 1~2 seconds)}
\usage{
fasta_parse_file(fasta_files, fasta_id_type = "uniprot")
}
\arguments{
\item{fasta_files}{array of fill paths to fasta files}

\item{fasta_id_type}{fasta type, typically this is 'uniprot'}
}
\description{
reasonably fast method for reading header rows from all fasta files (without over-complicating code/implementation to improve by a mere 1~2 seconds)
}
