% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{format_r_code}
\alias{format_r_code}
\title{placeholder title
parameter can either be a string or an array of strings (eg; lines from R history)}
\usage{
format_r_code(rcode)
}
\arguments{
\item{rcode}{todo}
}
\description{
placeholder title
parameter can either be a string or an array of strings (eg; lines from R history)
}
