% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{format_r_code}
\alias{format_r_code}
\title{format R code
parameter can either be a string or an array of strings (eg; lines from R history)}
\usage{
format_r_code(rcode_lines)
}
\arguments{
\item{rcode}{R code as an array of characters (strings)}
}
\description{
format R code
parameter can either be a string or an array of strings (eg; lines from R history)
}
