% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_rmarkdown.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Create the Quality Control report}
\usage{
generate_pdf_report(
  dataset,
  output_dir,
  norm_algorithm = "vwmb",
  pca_sample_labels = "auto"
)
}
\arguments{
\item{dataset}{a valid dataset object generated upstream by, for instance, import_dataset_skyline}

\item{output_dir}{output directory where all output files are stored, must be an existing directory}

\item{norm_algorithm}{normalization algorithm(s) used to normalize the QC proportion of data. options; "", "vsn", "loess", "rlr", "msempire", "vwmb", "modebetween". Provide an array of options to run each algorithm consecutively}

\item{pca_sample_labels}{see plot_sample_pca() function for params}
}
\description{
Normally used as part of the analysis_quickstart() function, refer to its implementation for example code.
}
