% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalization.R
\name{ggplot_coefficient_of_variation}
\alias{ggplot_coefficient_of_variation}
\title{placeholder title}
\usage{
ggplot_coefficient_of_variation(tib_input, samples, samples_colors)
}
\arguments{
\item{tib_input}{todo}

\item{samples}{todo}

\item{samples_colors}{todo}
}
\description{
placeholder title
}
