% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_encyclopedia.R
\name{import_dataset_encyclopedia}
\alias{import_dataset_encyclopedia}
\title{Import a label-free proteomics dataset from EncyclopeDIA}
\usage{
import_dataset_encyclopedia(
  file_quant_report,
  path_elib_rawfiles = NA,
  confidence_threshold = 0.01,
  return_decoys = FALSE
)
}
\arguments{
\item{file_quant_report}{the EncyclopeDIA 'Quant Report' .elib file}

\item{path_elib_rawfiles}{optional. the directory that contains the EncyclopeDIA .elib files matching each sample (mzML/dia input file to EncyclopeDIA)}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{return_decoys}{logical indicating whether to return decoy peptides. <not available for EncyclopeDIA at this point>}
}
\description{
This function imports peptide intensities, retention times and peptide-to-protein mappings from a 'Quant Report' .elib file generated by EncyclopeDIA.
}
\details{
Because this 'Quant Report' does not contain confidence scores per peptide per individual sample, these can optionally be imported from the individual .elib files generated by EncyclopeDIA while processing the individual samples / raw files.
This will enable MS-DAP to discriminate between match-between-runs hits and peptides that were detected/identified in a sample, and consequently use this information in peptide filtering (for instance by removing peptides that are mostly based on MBR hits).
}
