% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{import_dataset_fragpipe_ionquant}
\alias{import_dataset_fragpipe_ionquant}
\title{Import a label-free proteomics dataset from FragPipe presented in MSstats CSV format}
\usage{
import_dataset_fragpipe_ionquant(
  path,
  acquisition_mode,
  confidence_threshold = 0.01,
  collapse_peptide_by = "sequence_modified"
)
}
\arguments{
\item{path}{the full file path to the fragpipe output directory}

\item{acquisition_mode}{the type of experiment, should be a string. options: "dda" or "dia"}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}
}
\description{
To generate output files that we here require in MS-DAP, configure FragPipe as follows:
\itemize{
\item assign Experiment IDs in the workflow tab (you may simply set these all to 1)
\item enable IonQuant ("Quant (MS1)" tab, as of FragPipe v15)
\item optionally, enable match-between-runs
}
}
\details{
This function will merge data from various FragPipe output files to construct a MS-DAP dataset
}
