% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe_ionquant.R
\name{import_dataset_fragpipe_ionquant}
\alias{import_dataset_fragpipe_ionquant}
\title{Import a label-free proteomics dataset from FragPipe}
\usage{
import_dataset_fragpipe_ionquant(
  path,
  acquisition_mode,
  confidence_threshold = 0.01,
  collapse_peptide_by = "sequence_modified"
)
}
\arguments{
\item{path}{the full file path to the FragPipe output directory}

\item{acquisition_mode}{the type of experiment, should be a string. Valid options: "dda" or "dia"}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified', should be a numeric value between 0 and 1 (peptides with a 'Expectation' value in psm.tsv that is <= this threshold are classified as 'detected')}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}
}
\description{
To generate output files that we here require in MS-DAP, configure FragPipe as follows:
\enumerate{
\item assign unique Experiment identifiers in the FragPipe workflow tab.
Importantly, Fragpipe produces output (that MS-DAP can import) for each unique Experiment*Bioreplicate combination.
So if you provide the same experiment ID for all input files, the output looks like 1 sample to MS-DAP which is not what we want;
ensure all experiments have unique values in the FragPipe workflow tab !
\item enable IonQuant in the "Quant (MS1)" tab
(typically, this is already done if you loaded the LFQ-MBR workflow in the FragPipe workflow tab).
\item optionally, enable match-between-runs in the FragPipe "Quant (MS1)" tab.
}

This function will merge data from various FragPipe output files to construct a MS-DAP dataset.
If you followed above instructions, FragPipe should have generated the following files in the
output directory that will be parsed by this function;
\itemize{
\item combined_protein.tsv
\item for each experiment, a folder that includes 2 MS-DAP required files: ion.tsv and psm.tsv
}

(as most recently tested with FragPipe v18 and 19.1)
}
