% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fragpipe.R
\name{import_dataset_fragpipe_psm_file}
\alias{import_dataset_fragpipe_psm_file}
\title{Alternative FragPipe workflow, operates directly on a single psm.tsv file}
\usage{
import_dataset_fragpipe_psm_file(
  filename,
  intensity_sum = TRUE,
  acquisition_mode,
  confidence_threshold = 0.01,
  collapse_peptide_by = "sequence_modified"
)
}
\arguments{
\item{filename}{the full file path of the input file}

\item{intensity_sum}{boolean value whether to take the sum intensity of all PSM per precursor (TRUE, default) or use the value from the PSM with highest abundance (FALSE)}

\item{acquisition_mode}{the type of experiment, should be a string. options: "dda" or "dia"}

\item{confidence_threshold}{confidence score threshold at which a peptide is considered 'identified' (target value must be lesser than or equals)}

\item{collapse_peptide_by}{if multiple data points are available for a peptide in a sample, at what level should these be combined? options: "sequence_modified" (recommended default), "sequence_plain", ""}
}
\description{
For typical FragPipe workflows, \code{import_dataset_fragpipe_ionquant()} should be used instead of this function !

This function directly parses data from a psm.tsv file and doesn't use MBR data produced by FragPipe;
use this function for legacy FragPipe data  OR  FragPipe runs where no experiment info was provided at all in
the FragPipe workflow tab.

if intensity_sum is set to FALSE, the most abundant PSM per peptide*sample is selected to represent abundance.
}
