% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peptide_data.R
\name{import_protein_metadata_from_fasta}
\alias{import_protein_metadata_from_fasta}
\title{placeholder title}
\usage{
import_protein_metadata_from_fasta(
  protein_id,
  fasta_files,
  fasta_id_type = "uniprot",
  protein_separation_character = ";"
)
}
\arguments{
\item{protein_id}{todo}

\item{fasta_files}{todo}

\item{fasta_id_type}{todo}

\item{protein_separation_character}{todo}
}
\description{
placeholder title
}
