% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_files.R
\name{map_headers}
\alias{map_headers}
\title{placeholder title}
\usage{
map_headers(headers, l, error_on_missing = FALSE, allow_multiple = FALSE)
}
\arguments{
\item{headers}{todo}

\item{l}{list of mappings, where list names are target names and values are arrays of patterns that should match provided table headers}

\item{error_on_missing}{todo}

\item{allow_multiple}{todo}
}
\description{
placeholder title
}
