% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize a numeric matrix}
\usage{
normalize_matrix(x_as_log2, algorithm, mask_sample_groups = NA)
}
\arguments{
\item{x_as_log2}{the numeric matrix, all values must be log2 transformed}

\item{algorithm}{the normalization algorithm to apply. Either one of the built-in options, see normalization_algorithms(), or the name of your custom normalization function (see online documentation/examples)}

\item{mask_sample_groups}{a character array of the same length as the number of columns in the provided data matrix, describing the grouping of each column/sample}
}
\description{
Normalize a numeric matrix
}
