% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_normalization.R
\name{normalize_msempire}
\alias{normalize_msempire}
\title{adapted from msEmpiRe::normalize(), supports any input matrix + definition of groups
input expected as finite values, not log transformed}
\usage{
normalize_msempire(x, mask_sample_groups)
}
\arguments{
\item{x}{todo}

\item{mask_sample_groups}{todo}
}
\description{
adapted from msEmpiRe::normalize(), supports any input matrix + definition of groups
input expected as finite values, not log transformed
}
