% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{peptide_and_protein_counts_per_sample}
\alias{peptide_and_protein_counts_per_sample}
\title{add peptide and protein counts to samples tibble, so downstream code includes this in output tables and report figures relating to sample metadata}
\usage{
peptide_and_protein_counts_per_sample(peptides, samples, isdia)
}
\arguments{
\item{peptides}{peptide tibble in long format}

\item{samples}{samples tibble in wide format}

\item{isdia}{boolean indicating this is a DIA dataset (DDA otherwise). Use is_dia_dataset(dataset) to obtain this from data}
}
\description{
add peptide and protein counts to samples tibble, so downstream code includes this in output tables and report figures relating to sample metadata
}
