% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_detect.R
\name{plot_differential_detect}
\alias{plot_differential_detect}
\title{placeholder title}
\usage{
plot_differential_detect(dataset)
}
\arguments{
\item{dataset}{dataset object}
}
\description{
placeholder title
}
