% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_data.R
\name{plot_peptide_data}
\alias{plot_peptide_data}
\title{plot peptide-level data}
\usage{
plot_peptide_data(
  dataset,
  select_all_proteins = FALSE,
  select_diffdetect_candidates = TRUE,
  select_dea_signif = TRUE,
  filter_protein_ids = NA,
  filter_genes = NA,
  output_dir,
  show_unused_datapoints = FALSE,
  norm_algorithm = NA
)
}
\arguments{
\item{dataset}{todo}

\item{select_all_proteins}{todo}

\item{select_diffdetect_candidates}{todo}

\item{select_dea_signif}{todo}

\item{filter_protein_ids}{todo}

\item{filter_genes}{todo}

\item{output_dir}{todo}

\item{show_unused_datapoints}{todo}

\item{norm_algorithm}{todo}
}
\description{
plot peptide-level data
}
