% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_pca.R
\name{plot_sample_pca}
\alias{plot_sample_pca}
\title{placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))}
\usage{
plot_sample_pca(
  matrix_sample_intensities,
  samples,
  samples_colors,
  sample_label_property = "auto",
  pch_as_exclude = TRUE,
  infer_continuous_scale = TRUE
)
}
\arguments{
\item{matrix_sample_intensities}{peptide or protein abundance matrix}

\item{samples}{sample metadata table}

\item{samples_colors}{sample colors in wide format tibble}

\item{sample_label_property}{labels used to represent the samples. Set to "auto" to automatically select based on number of samples (default). possible values; auto, shortname, index, index_asis}

\item{pch_as_exclude}{should exclude samples be represented with a distinct symbol shape? possible values; FALSE, TRUE (default)}

\item{infer_continuous_scale}{should the legends infer continuous scale from the data? FALSE = all categorical, TRUE = automatic (default)}
}
\description{
placeholder title
for reference, standard pca (does not cope with missing values); summary(stats::prcomp(t(matrix_sample_intensities), center = T, scale. = F))
}
