% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_expressionset.R
\name{protein_eset_from_data}
\alias{protein_eset_from_data}
\title{Convert a peptide-level ExpressionSet to protein-level by summing their respective peptide intensities per sample}
\usage{
protein_eset_from_data(x, eset)
}
\arguments{
\item{eset_peptides}{A Biobase ExpressionSet with peptide data. Intensity values are assumed to be log2 transformed! fData() must contain protein_id attributes}

\item{mode}{roll-up strategy}
}
\description{
Convert a peptide-level ExpressionSet to protein-level by summing their respective peptide intensities per sample
}
