% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup_pep2prot}
\alias{rollup_pep2prot}
\title{wrapper function for peptide-to-protein rollup}
\usage{
rollup_pep2prot(
  tib,
  intensity_is_log2 = TRUE,
  algo_rollup,
  return_as_matrix = FALSE
)
}
\arguments{
\item{tib}{peptide tibble in long-format. missing values should be set NA, not zero. required columns: peptide_id, protein_id, sample_id, intensity}

\item{intensity_is_log2}{boolean indicating log2 scaling}

\item{algo_rollup}{type of rollup to perform. options: maxlfq or maxlfq_iq for MaxLFQ rollup using iq package, maxlfq_diann for MaxLFQ rollup using diann package. Any other value performs rollup by peptide summation.}

\item{return_as_matrix}{instead of long-format tibble, return a protein*sample matrix (missing values set to NA)}
}
\description{
wrapper function for peptide-to-protein rollup
}
