% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata_from_file}
\alias{sample_metadata_from_file}
\title{placeholder title}
\usage{
sample_metadata_from_file(sample_id, filename, group_order = NA)
}
\arguments{
\item{sample_id}{todo}

\item{filename}{todo}

\item{group_order}{todo}
}
\description{
placeholder title
}
