% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata_from_filenames}
\alias{sample_metadata_from_filenames}
\title{placeholder title}
\usage{
sample_metadata_from_filenames(
  sample_id,
  sample_property_regex = list(),
  sample_exclude_regex = "",
  group_order = NA,
  group_regex_array = NA
)
}
\arguments{
\item{sample_id}{todo}

\item{sample_property_regex}{todo}

\item{sample_exclude_regex}{todo}

\item{group_order}{todo}

\item{group_regex_array}{todo}
}
\description{
placeholder title
}
