% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata_sort_and_filter}
\alias{sample_metadata_sort_and_filter}
\title{placeholder title}
\usage{
sample_metadata_sort_and_filter(
  df,
  sample_exclude_regex = "",
  group_order = NA
)
}
\arguments{
\item{df}{todo}

\item{sample_exclude_regex}{todo}

\item{group_order}{todo}
}
\description{
placeholder title
}
