% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{smooth_loess_custom}
\alias{smooth_loess_custom}
\title{wrapper for; predict(loess(y ~ x, span = span, na.action = na.omit), x)}
\usage{
smooth_loess_custom(x, y, span = 0.1, remove_na_yval = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{span}{argument for loess()}

\item{remove_na_yval}{set all output values where is.na(y) to NA}
}
\description{
also adds option to disregard predicted values if input y is NA.
does nothing if less than 10 values are passed.
}
