% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{split_array}
\alias{split_array}
\title{split array into a list of N-sized chunks}
\usage{
split_array(x, chunk_size)
}
\arguments{
\item{x}{array to be chunked}

\item{chunk_size}{integer, maximum chunk size}
}
\description{
split array into a list of N-sized chunks
}
