% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{strip_common_substring}
\alias{strip_common_substring}
\title{remove starting or trailing substrings that ar the same for all strings in a vector. eg; .txt in all filesnames, or <experiment data><samplename>
strip_common_substring(s = unique(peptides$sample_id))
strip_common_substring(s = c("sample_a.txt","sample_b1.txt","sample_xyz.txt"))
strip_common_substring(s = c("","sample_b1.txt","sample_xyz.txt"))
strip_common_substring(s = c("sample.txt","sample_b1.txt","sample_xyz.txt"))
assertthat::are_equal(strip_common_substring(c("sample_a.txt","sample_b1.txt","sample_xyz.txt"), c("a","b1","xyz")))}
\usage{
strip_common_substring(s)
}
\arguments{
\item{s}{todo}
}
\description{
remove starting or trailing substrings that ar the same for all strings in a vector. eg; .txt in all filesnames, or <experiment data><samplename>
strip_common_substring(s = unique(peptides$sample_id))
strip_common_substring(s = c("sample_a.txt","sample_b1.txt","sample_xyz.txt"))
strip_common_substring(s = c("","sample_b1.txt","sample_xyz.txt"))
strip_common_substring(s = c("sample.txt","sample_b1.txt","sample_xyz.txt"))
assertthat::are_equal(strip_common_substring(c("sample_a.txt","sample_b1.txt","sample_xyz.txt"), c("a","b1","xyz")))
}
