% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{summarise_stats__for_contrast}
\alias{summarise_stats__for_contrast}
\title{worker function for actual stats summary}
\usage{
summarise_stats__for_contrast(
  dataset,
  return_dea,
  return_diffdetect,
  contr,
  dea_algorithm,
  dea_logfc_as_effectsize,
  diffdetect_zscore_threshold,
  remove_ambiguous_proteingroups
)
}
\arguments{
\item{dataset}{see \code{summarise_stats()}}

\item{return_dea}{see \code{summarise_stats()}}

\item{return_diffdetect}{see \code{summarise_stats()}}

\item{contr}{contrast for which DD and/or DEA was performed (i.e. valid value in dataset$dd_proteins and dataset$de_proteins)}

\item{dea_algorithm}{DEA algorithm as used in \code{dea()} upstream}

\item{dea_logfc_as_effectsize}{see \code{summarise_stats()}}

\item{diffdetect_zscore_threshold}{see \code{summarise_stats()}}

\item{remove_ambiguous_proteingroups}{see \code{summarise_stats()}}
}
\description{
worker function for actual stats summary
}
