% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generic.R
\name{threshold_numerics}
\alias{threshold_numerics}
\title{basic threshold function}
\usage{
threshold_numerics(x, threshold)
}
\arguments{
\item{x}{numeric array or matrix}

\item{threshold}{limit values below this val}
}
\description{
basic threshold function
}
