% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_as_excel.R
\name{write_peptide_abundance_matrix_to_file}
\alias{write_peptide_abundance_matrix_to_file}
\title{Create an Excel document with peptide and protein abundances}
\usage{
write_peptide_abundance_matrix_to_file(dataset, filename, norm_algorithm = "")
}
\arguments{
\item{dataset}{your dataset}

\item{filename}{full file path to output file}

\item{norm_algorithm}{normalization algorithm(s)}
}
\description{
Conditional formatting is used to highlight match-between-runs peptides (low confidence data points).
}
\details{
Keep in mind that large datasets may result in (too) large Excel files, so in some cases you're better of pre-processing your data in R.

included tables;
- peptide intensity (-filter +norm. conditional formatting)
- peptide identification (1/0 flags)
- peptide intensity_all_group (+filter +norm)
- protein roll-up from peptide intensity
- protein roll-up from peptide intensity_all_group
}
